% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_annotations}
\alias{add_annotations}
\title{Add an annotation(s) to a plot}
\usage{
add_annotations(p, text = NULL, ..., data = NULL)
}
\arguments{
\item{p}{a plotly object}

\item{text}{annotation text (required).}

\item{...}{these arguments are documented at \url{https://plot.ly/r/reference/#layout-annotations}}

\item{data}{a data frame.}
}
\description{
Add an annotation(s) to a plot
}
\examples{

# single annotation
plot_ly(mtcars, x = ~wt, y = ~mpg) \%>\%
  slice(which.max(mpg)) \%>\%
  annotation(text = "Good mileage")
  
# multiple annotations
plot_ly(mtcars, x = ~wt, y = ~mpg) \%>\%
  filter(gear == 5) \%>\%
  annotation("five cylinder", ax = 40) 
  
}
\author{
Carson Sievert
}

