% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_lines}
\alias{add_lines}
\title{Add lines to a plotly vis}
\usage{
add_lines(p, ...)
}
\arguments{
\item{p}{a plotly or ggplot object.}

\item{...}{These arguments are documented in the references section below.
Note that acceptable arguments depend on the trace type.}
}
\description{
Equivalent to \code{add_paths}, but with the x-values sorted.
}

