% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_markers}
\alias{add_markers}
\title{Add points to a plotly vis}
\usage{
add_markers(p, ...)
}
\arguments{
\item{p}{a plotly or ggplot object.}

\item{...}{These arguments are documented in the references section below.
Note that acceptable arguments depend on the trace type.}
}
\description{
Add points to a plotly vis
}

