% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_paths}
\alias{add_paths}
\title{Add paths to a plotly vis}
\usage{
add_paths(p, ...)
}
\arguments{
\item{p}{a plotly or ggplot object.}

\item{...}{These arguments are documented in the references section below.
Note that acceptable arguments depend on the trace type.}
}
\description{
Add paths to a plotly vis
}

