% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_polygons}
\alias{add_polygons}
\title{Add polygons to a plotly vis}
\usage{
add_polygons(p, ...)
}
\arguments{
\item{p}{a plotly or ggplot object.}

\item{...}{These arguments are documented in the references section below.
Note that acceptable arguments depend on the trace type.}
}
\description{
Add polygons to a plotly vis
}
\examples{

library(dplyr)
data(canada.cities, package = "maps")

ggplot2::map_data("world", "canada") \%>\%
  group_by(group) \%>\%
  plot_ly(x = ~long, y = ~lat, hoverinfo = "none") \%>\%
  add_markers(text = ~paste(name, "<br />", pop), hoverinfo = "text",
    data = canada.cities) \%>\%
  layout(showlegend = FALSE)
}

