% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_ribbons}
\alias{add_ribbons}
\title{Add ribbons to a plotly vis}
\usage{
add_ribbons(p, ...)

add_ribbons(p, ...)
}
\arguments{
\item{p}{a plotly or ggplot object.}

\item{...}{These arguments are documented in the references section below.
Note that acceptable arguments depend on the trace type.}
}
\description{
Ribbons are a special case of polygons.

Ribbons are a special case of polygons.
}

