% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_text}
\alias{add_text}
\title{Add text to a plotly vis}
\usage{
add_text(p, text = NULL, ...)
}
\arguments{
\item{p}{a plotly or ggplot object.}

\item{...}{These arguments are documented in the references section below.
Note that acceptable arguments depend on the trace type.}
}
\description{
Add text to a plotly vis
}

