% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotly.R
\name{add_trace}
\alias{add_trace}
\title{Add a trace to a plotly object}
\usage{
add_trace(data = data.frame, ...)
}
\arguments{
\item{data}{A data frame with a class of plotly.}

\item{...}{Trace arguments. Arguments are evaluated in the environment attached to
the most recent trace. See the reference below for documentation.}
}
\description{
Turns a dataset into a plotly object. A plotly object can be conceptualized as
a set of traces, a layout. This function will initiate
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#Trace_objects}
}

