\docType{data}
\name{aes2marker}
\alias{aes2marker}
\title{Convert ggplot2 aes to plotly "marker" codes.}
\format{\preformatted{ Named chr [1:6] "opacity" "symbol" "color" "size" "symbol" ...
 - attr(*, "names")= chr [1:6] "alpha" "pch" "colour" "size" ...
}}
\usage{
aes2marker
}
\description{
Convert ggplot2 aes to plotly "marker" codes.
}
\keyword{datasets}

