% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{animate}
\alias{animate}
\title{Animate a collection of plots}
\usage{
animate(..., slider = FALSE, button = TRUE)
}
\arguments{
\item{...}{a collection of ggplot2 or plotly objects}

\item{slider}{whether to populate a slider to control animation}

\item{button}{whether to populate a play button (recommended).}
}
\description{
Animate a collection of plots
}
\examples{

p <- plot_ly(x = LETTERS, y = seq_along(LETTERS), color = LETTERS) \%>\% hide_legend()
p <- plot_ly(x = LETTERS, y = seq_along(LETTERS))# \%>\% hide_legend()
a <- animate(
  add_markers(p), add_bars(p)#, 
  #plot_ly(labels = LETTERS, values = seq_along(LETTERS)) \%>\% add_pie()
)
# TODO: how to optionally suppress the slider?
animation_opts(a, 1000, redraw = TRUE)


}
\author{
Carson Sievert
}
