% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{animation_button}
\alias{animation_button}
\title{Hide or customize the animation button}
\usage{
animation_button(p, ...)
}
\arguments{
\item{p}{a plotly object}

\item{...}{arguments passed to the updatemenus which controls the play
button \url{https://github.com/plotly/plotly.js/blob/master/src/components/updatemenus/attributes.js}}
}
\description{
Hide or customize the animation button
}
\author{
Carson Sievert
}
\seealso{
\code{\link{animation_opts}()}, \code{\link{animation_button}()}
}

