% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{animation_slider}
\alias{animation_slider}
\title{Hide or customize the animation button}
\usage{
animation_slider(p, hide = FALSE, ...)
}
\arguments{
\item{p}{a plotly object}

\item{hide}{remove the animation slider?}

\item{...}{attributes passed to the sliders object which controls the animation
slider \url{https://github.com/plotly/plotly.js/blob/master/src/components/sliders/attributes.js}}
}
\description{
Hide or customize the animation button
}
\author{
Carson Sievert
}
\seealso{
\code{\link{animation_opts}()}, \code{\link{animation_button}()}
}

