% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{animationOpts}
\alias{animationOpts}
\title{Animation options}
\usage{
animationOpts(frameDuration = 500, transitionDuration = 500,
  easing = "cubic-in-out", redraw = FALSE)
}
\arguments{
\item{frameDuration}{The duration each frame is shown (in milliseconds).}

\item{transitionDuration}{The duration of the transition (in milliseconds).}

\item{easing}{The type of transition easing. See the list of options here
\url{https://github.com/plotly/plotly.js/blob/master/src/plots/animation_attributes.js}}

\item{redraw}{Trigger a redraw of the plot at completion of the transition?
A redraw may significantly impact performance, but may be necessary to 
update plot attributes that can't be transitioned.}
}
\description{
Animation options
}
\author{
Carson Sievert
}
\seealso{
\code{\link{animate}()}
}

