% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/layout.R
\name{annotation}
\alias{annotation}
\title{Create an annotation object.}
\usage{
annotation(x = NULL, y = NULL, xref = NULL, yref = NULL, text = NULL,
  showarrow = NULL, font = NULL, xanchor = NULL, yanchor = NULL,
  align = NULL, arrowhead = NULL, arrowsize = NULL, arrowwidth = NULL,
  arrowcolor = NULL, ax = NULL, ay = NULL, textangle = NULL,
  bordercolor = NULL, borderwidth = NULL, borderpad = NULL,
  bgcolor = NULL, opacity = NULL)
}
\arguments{
\item{x,}{y, xref, yref, text, showarrow, font, xanchor, yanchor, align,
arrowhead, arrowsize, arrowwidth, arrowcolor, ax, ay, textangle, bordercolor,
borderwidth, borderpad, bgcolor, opacity}
}
\description{
Create an annotation object.
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#annotation}
}

