% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trace.R
\name{area}
\alias{area}
\title{Create a 'area' trace object}
\usage{
area(r = NULL, t = NULL, name = NULL, marker = NULL,
  showlegend = NULL, stream = NULL, visible = NULL, angularaxis = NULL,
  radialaxis = NULL)
}
\arguments{
\item{r,}{t, name, marker, showlegend, stream, visible, angularaxis,
radialaxis}
}
\description{
Create a 'area' trace object
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#area}
}

