% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{as_widget}
\alias{as_widget}
\title{Convert a list to a plotly htmlwidget object}
\usage{
as_widget(x, ...)
}
\arguments{
\item{x}{a plotly object.}

\item{...}{other options passed onto \code{htmlwidgets::createWidget}}
}
\description{
Convert a list to a plotly htmlwidget object
}
\examples{

trace <- list(x = 1, y = 1)
obj <- list(data = list(trace), layout = list(title = "my plot"))
as_widget(obj)

}
