% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{attrs_selected}
\alias{attrs_selected}
\title{Specify attributes of selection traces}
\usage{
attrs_selected(opacity = 1, ...)
}
\arguments{
\item{opacity}{a number between 0 and 1 specifying the overall opacity of
the selected trace}

\item{...}{other trace attributes attached to the selection trace.}
}
\description{
By default the name of the selection trace derives from the selected values.
}
\author{
Carson Sievert
}
