% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trace.R
\name{bar}
\alias{bar}
\title{Create a 'bar' trace object}
\usage{
bar(x = NULL, y = NULL, name = NULL, orientation = NULL, text = NULL,
  error_y = NULL, error_x = NULL, marker = NULL, opacity = NULL,
  xaxis = NULL, yaxis = NULL, showlegend = NULL, stream = NULL,
  visible = NULL, xsrc = NULL, ysrc = NULL, r = NULL, t = NULL)
}
\arguments{
\item{x,}{y, name, orientation, text, error_y, error_x, marker, opacity,
xaxis, yaxis, showlegend, stream, visible, xsrc, ysrc, r, t}
}
\description{
Create a 'bar' trace object
}
\examples{
\dontrun{
x <- with(diamonds, tapply(price, cut, mean))
bar(names(x), as.numeric(x))
# TODO: an example of layering bar traces to get a mosiac chart?
}
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#bar}
}

