% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{bind}
\alias{bind}
\title{Bind an event to a plotly htmlwidget}
\usage{
bind(p, event = "plotly_click", callback = "console.log(event, data)")
}
\arguments{
\item{p}{a plotly visualization}

\item{event}{A valid DOM event}

\item{callback}{character string containing the body of a JavaScript 
callback function with the arguments \code{event} and \code{data} 
to be applied to the plotly instance.}
}
\description{
This functionality is an experiment
}
\examples{

p <- plot_ly(mtcars, x = mpg, y = disp, mode = "markers")
# open this in a web browser, click on a point, and open your console
bind(p)
# access the clicked point and display alert box
cb <- c(
  "var pts = data.points[0];",
  "alert('You selected ['+pts.x+', '+pts.y+']');"
 )
bind(p, callback = cb)

w1 <- dplyr::count(mtcars, cyl) \%>\%
  plot_ly(x = cyl, y = n, type = "bar") \%>\%
  bind(callback = "") \%>\%
  as.widget()
  
 w2 <- mtcars \%>\%
   plot_ly(x = mpg, type = "hist") \%>\%
   as.widget()

library(htmltools)
browsable(tagList(w1, w2))



}

