% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trace.R
\name{box}
\alias{box}
\title{Create a 'box' trace object}
\usage{
box(y = NULL, x0 = NULL, x = NULL, name = NULL, boxmean = NULL,
  boxpoints = NULL, jitter = NULL, pointpos = NULL, whiskerwidth = NULL,
  fillcolor = NULL, marker = NULL, line = NULL, opacity = NULL,
  xaxis = NULL, yaxis = NULL, showlegend = NULL, stream = NULL,
  visible = NULL, xsrc = NULL, ysrc = NULL)
}
\arguments{
\item{y,}{x0, x, name, boxmean, boxpoints, jitter, pointpos, whiskerwidth,
fillcolor, marker, line, opacity, xaxis, yaxis, showlegend, stream, visible,
xsrc, ysrc}
}
\description{
Create a 'box' trace object
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#box}
}

