% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_table.R
\name{build_table}
\alias{build_table}
\title{Build a ggplot2/plotly comparison table}
\usage{
build_table(p1, gg1, p2 = "ropensci/plotly", gg2, width = 400,
  height = 300, ...)
}
\arguments{
\item{p1}{New version of plotly.}

\item{gg1}{The ggplot2 dependency for \code{p1}. If not specified, the currently
installed version is used as a dependency.}

\item{p2}{Old version of plotly (should be master).}

\item{gg2}{The ggplot2 dependency for \code{p2}.}

\item{width}{width of thumbnails (in pixels)}

\item{height}{height of thumbnails (in pixels)}
}
\description{
This function is mainly for development purposes, but users might also find
it useful to see the capabilities/limitations of \link{ggplotly}
}
\details{
If \code{package_version} is used to specify a version,
this will try to install the corresponding CRAN version via
\code{devtools::install_version}. Otherwise, versions are passed onto
the \code{repo} argument in \code{devtools::install_github}.
}
\examples{
\dontrun{
# by default, use latest CRAN versions?
build_table()

# compare #269 (with dev of version of ggplot2) versus plotly master with ggplot2 1.0.1 from CRAN
build_table(
  "ropensci/plotly#269",
  "hadley/ggplot2",
  gg2 = package_version("1.0.1")
)

# idea a pipeable syntax for building a table?
init_table(width, height, ...) \%>\%
  add_column("ropensci/plotly")


}
}

