% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{colorbar}
\alias{colorbar}
\title{Modify colorbar}
\usage{
colorbar(p, ...)
}
\arguments{
\item{p}{a plotly object}

\item{...}{arguments are documented here 
\url{https://plot.ly/r/reference/#scatter-marker-colorbar}.}
}
\description{
Modify colorbar
}
\examples{

}
\author{
Carson Sievert
}

