% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{config}
\alias{config}
\title{Set the default configuration for plotly}
\usage{
config(p, ..., collaborate = TRUE, cloud = FALSE, locale = NULL)
}
\arguments{
\item{p}{a plotly object}

\item{...}{these arguments are documented at
\url{https://github.com/plotly/plotly.js/blob/master/src/plot_api/plot_config.js}}

\item{collaborate}{include the collaborate mode bar button (unique to the R pkg)?}

\item{cloud}{include the send data to cloud button?}

\item{locale}{locale to use. See \href{https://github.com/plotly/plotly.js/tree/master/dist#to-include-localization}{here} for more info}
}
\description{
Set the default configuration for plotly
}
\examples{

p <- plot_ly(x = rnorm(1000))

config(p, displaylogo = FALSE, collaborate = FALSE)
config(p, locale = "de")
}
\author{
Carson Sievert
}
