% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trace.R
\name{contour}
\alias{contour}
\title{Create a 'contour' trace object}
\usage{
contour(z = NULL, x = NULL, y = NULL, name = NULL, zauto = NULL,
  zmin = NULL, zmax = NULL, autocontour = NULL, ncontours = NULL,
  contours = NULL, line = NULL, colorscale = NULL, reversescale = NULL,
  showscale = NULL, colorbar = NULL, connectgaps = NULL, opacity = NULL,
  xaxis = NULL, yaxis = NULL, showlegend = NULL, stream = NULL,
  visible = NULL, x0 = NULL, dx = NULL, y0 = NULL, dy = NULL,
  xtype = NULL, ytype = NULL)
}
\arguments{
\item{z,}{x, y, name, zauto, zmin, zmax, autocontour, ncontours, contours,
line, colorscale, reversescale, showscale, colorbar, connectgaps, opacity,
xaxis, yaxis, showlegend, stream, visible, x0, dx, y0, dy, xtype, ytype}
}
\description{
Create a 'contour' trace object
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#contour}
}

