% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aux.R
\name{contours}
\alias{contours}
\title{Create a 'contours' auxiliary object}
\usage{
contours(showlines = NULL, start = NULL, end = NULL, size = NULL,
  coloring = NULL)
}
\arguments{
\item{showlines,}{start, end, size, coloring}
}
\description{
Available in \link{contour} and \link{histogram2dcontour}.
}
\examples{
\dontrun{

}
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#contours}
}

