% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstalk.R
\name{crosstalk}
\alias{crosstalk}
\title{Modify crosstalk selection options}
\usage{
crosstalk(p, on = "plotly_selected", off = "plotly_deselect",
  color = NULL, dynamic = FALSE, persistent = FALSE, opacityDim = 0.2,
  showInLegend = FALSE)
}
\arguments{
\item{p}{a plotly visualization.}

\item{on}{turn on a selection on which event(s)? Likely candidates are
'plotly_hover', 'plotly_click', 'plotly_selected'.}

\item{off}{turn off a selection on which event(s)? Likely candidates are
'plotly_unhover', 'plotly_doubleclick', 'plotly_deselect'.}

\item{color}{character string of color(s) to use for 
highlighting selections. See \code{\link{toRGB}()} for valid color
specifications. If \code{NULL} (the default), the color of selected marks
are not altered (only their opacity).}

\item{dynamic}{should UI controls for managing selection aesthetics be 
included in the output?}

\item{persistent}{should selections persist (i.e., accumulate)?}

\item{opacityDim}{a number between 0 and 1 used to reduce the
opacity of non-selected traces (by multiplying with the existing opacity).}

\item{showInLegend}{populate an additional legend entry for the selection?}
}
\description{
Control the visual appearance of selections deriving from a given
selection set.
}

