% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstalk.R
\name{ct_opts}
\alias{ct_opts}
\title{Control selection appearance}
\usage{
ct_opts(dynamic = FALSE, persitant = dynamic, color = NULL,
  opacityDim = 0.2, showInLegend = FALSE)
}
\arguments{
\item{dynamic}{should UI controls for managing selection aesthetics be 
included in the output?}

\item{color}{color(s) to use for highlighting selections. 
If \code{NULL} (the default), the color of selections are not altered. 
If not \code{NULL}
a valid color code,  
If not \code{dynamic}, this argument should be length 1,
If \code{dynamic},
this argument accepts a character}

\item{opacityDim}{a number between 0 and 1 used to reduce the
opacity of non-selected traces (by multiplying with the existing opacity).}

\item{showInLegend}{show a legend entry for additional "selection traces"?}

\item{persistant}{should selections persist (i.e., accumulate)?}
}
\description{
Control the visual appearance of selections deriving from a given
selection set.
}

