% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{embed_notebook}
\alias{embed_notebook}
\title{Embed a plotly iframe into a IPython Notebook}
\usage{
embed_notebook(x, width = "100\%", height = "525")
}
\arguments{
\item{x}{a plotly object}

\item{width}{attribute of the iframe}

\item{height}{attribute of the iframe}
}
\description{
Embed a plotly iframe into a IPython Notebook
}

