% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{ensure_file_exist}
\alias{ensure_file_exist}
\title{Create file if nonexistent}
\usage{
ensure_file_exist(abspath)
}
\arguments{
\item{abspath}{Character vector of file path}
}
\description{
Create file if nonexistent
}

