% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aux.R
\name{error_y}
\alias{error_y}
\title{Create an 'error_y' auxiliary object}
\usage{
error_y(type = NULL, symmetric = NULL, array = NULL, value = NULL,
  arrayminus = NULL, valueminus = NULL, color = NULL, thickness = NULL,
  width = NULL, opacity = NULL, visible = NULL)
}
\arguments{
\item{type,}{symmetric, array, value, arrayminus, valueminus, color,
thickness, width, opacity, visible}
}
\description{
Available in \link{scatter}, \link{bar}, \link{histogram}, and \link{scatter3d}.
}
\examples{
\dontrun{
means <- with(diamonds, tapply(carat, cut, mean))
sds <- with(diamonds, tapply(carat, cut, sd))
# construct the error bar object
err <- error_y(array = as.numeric(sds))
# pass it to a trace object
scatter(names(means), as.numeric(means), error_y = err)
}
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#error_y}
}

