% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aux.R
\name{error_z}
\alias{error_z}
\title{Create a 'error_z' auxiliary object}
\usage{
error_z(type = NULL, symmetric = NULL, array = NULL, value = NULL,
  arrayminus = NULL, valueminus = NULL, color = NULL, thickness = NULL,
  width = NULL, opacity = NULL, visible = NULL)
}
\arguments{
\item{type,}{symmetric, array, value, arrayminus, valueminus, color,
thickness, width, opacity, visible.}
}
\description{
Available in \link{scatter3d}.
}
\examples{
\dontrun{

}
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#error_z}
}

