% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{event_data}
\alias{event_data}
\title{Access plotly user input event data in shiny}
\usage{
event_data(event = c("plotly_click", "plotly_selected"), source = "A")
}
\arguments{
\item{event}{The type of plotly event. Currently 'plotly_click' and 
'plotly_selected' are supported.}

\item{source}{Which plot should the listener be tied to? This 
(character string) should match the value of \code{source} in \link{plot_ly}.}
}
\description{
This function must be called within a reactive shiny context.
}
\examples{
\dontrun{
shiny::runApp(system.file("examples", "events", package = "plotly"))
}
}
\author{
Carson Sievert
}

