% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{event_register}
\alias{event_register}
\title{Register a shiny input value}
\usage{
event_register(p, event = NULL)
}
\arguments{
\item{p}{a plotly object.}

\item{event}{The type of plotly event. All supported events are listed in the
function signature above (i.e., the usage section).}
}
\description{
Register a shiny input value
}
\seealso{
\link{event_data}
}
\author{
Carson Sievert
}
