% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{event_unregister}
\alias{event_unregister}
\title{Un-register a shiny input value}
\usage{
event_unregister(p, event = NULL)
}
\arguments{
\item{p}{a plotly object.}

\item{event}{The type of plotly event. All supported events are listed in the
function signature above (i.e., the usage section).}
}
\description{
Un-register a shiny input value
}
\seealso{
\link{event_data}
}
\author{
Carson Sievert
}
