% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{explain_attrs}
\alias{explain_attrs}
\title{Show information about attribute(s) of a given trace type}
\usage{
explain_attrs(type = "scatter", attrs = c("x", "y"))
}
\arguments{
\item{type}{a trace type}

\item{attrs}{attributes for the race type}
}
\description{
Show information about attribute(s) of a given trace type
}
\examples{

explain_attrs()
}

