% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/layout.R
\name{font}
\alias{font}
\title{Create a font object.}
\usage{
font(family = NULL, size = NULL, color = NULL, outlinecolor = NULL)
}
\arguments{
\item{family,}{size, color, outlinecolor}
}
\description{
Available as: \itemize{
\item \code{tickfont} in \link{colorbar}, \link{xaxis}, \link{yaxis}, \link{zaxis}
\item \code{textfont} in \link{scatter}
\item \code{font} in \link{layout}, \link{legend}, \link{annotation}
\item \code{titlefont} in \link{colorbar}, \link{xaxis}, \link{yaxis}, \link{zaxis}, \link{layout}
}
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#font}
}

