% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{geo}
\alias{geo}
\title{Initiate a plotly-geo object}
\usage{
geo(data = data.frame(), ...)
}
\arguments{
\item{...}{arguments passed along to \code{\link{plot_ly}()}.}
}
\description{
Use this function instead of \code{\link{plot_ly}()} to initialize
a plotly-geo object. This enforces the entire plot so use
the scattergeo trace type, and enables higher level geometries
like \code{\link{add_polygons}()} to work
}
\examples{

map_data("world", "canada") \%>\%
  group_by(group) \%>\%
  geo(x = ~lat, y = ~long) \%>\%
  add_polygons()

}

