\docType{data}
\name{geom2trace}
\alias{geom2trace}
\title{Convert basic geoms to traces.}
\format{\preformatted{List of 3
 $ path   :function (data, params)  
 $ polygon:function (data, params)  
 $ point  :function (data, params)  
}}
\usage{
geom2trace
}
\description{
Convert basic geoms to traces.
}
\keyword{datasets}

