% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_boxplot2.R
\name{geom_boxplot2}
\alias{geom_boxplot2}
\title{Attempt to convert \code{geom_boxplot()} to a plotly.js box trace}
\usage{
geom_boxplot2(...)
}
\arguments{
\item{...}{arguments passed along to \code{\link[ggplot2:geom_boxplot]{ggplot2::geom_boxplot()}}}
}
\description{
There are two ways to create boxplot via \code{\link[=ggplotly]{ggplotly()}}: with either
this function or \code{\link[ggplot2:geom_boxplot]{ggplot2::geom_boxplot()}}. This function uses
the \href{https://plot.ly/r/reference/#box}{box} trace type whereas the
latter uses a combination of \href{https://plot.ly/r/reference/#scatter}{scatter}
traces to render the visualization. This implies that, this
function lets plotly.js compute boxplot summaries and positional
dodging, whereas the latter uses the actual ggplot2 boxplot
definition(s).
}
\examples{

subplot(
  ggplot(diamonds) + geom_boxplot(aes(y = price)),
  ggplot(diamonds) + geom_boxplot2(aes(y = price))
)
}
