% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tools.R
\name{get_config_file}
\alias{get_config_file}
\title{Read Plotly config file (which is a JSON) and create one if nonexistent}
\usage{
get_config_file(args = c())
}
\arguments{
\item{args}{Character vector of keys you are looking up}
}
\value{
List of keyword-value pairs (config)
}
\description{
Read Plotly config file (which is a JSON) and create one if nonexistent
}
\examples{
\dontrun{
get_config_file(c("plotly_domain", "plotly_streaming_domain"))
}
}

