% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotly.R
\name{get_figure}
\alias{get_figure}
\title{Request data/layout for a particular Plotly figure}
\usage{
get_figure(username, id)
}
\arguments{
\item{username}{corresponding username for the figure.}

\item{id}{of the Plotly figure.}
}
\description{
Request data/layout for a particular Plotly figure
}
\examples{
\dontrun{
 # https://plot.ly/~TestBot/100
 resp <- get_figure("TestBot", "100")
 names(resp[["layout"]])
 names(resp[["data"]])
}
}
\references{
https://plot.ly/rest/
}

