% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{get_plot}
\alias{get_plot}
\title{Obtain underlying data of plotly object}
\usage{
get_plot(data, strict = TRUE)
}
\arguments{
\item{data}{a data frame with a class of plotly (and a plotly_hash attribute).}

\item{srict}{throw a warning if the plotly_hash attribute is missing.}
}
\description{
Obtain underlying data of plotly object
}

