% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggridges.R
\name{get_ridge_data}
\alias{get_ridge_data}
\title{Get data for ridge plots}
\usage{
get_ridge_data(data, na.rm)
}
\arguments{
\item{data}{dataframe, the data returned by \code{ggplot2::ggplot_build()}.}

\item{na.rm}{boolean, from params}
}
\value{
dataframe containing plotting data
}
\description{
Get data for ridge plots
}
