% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{get_x}
\alias{get_x}
\title{Obtain x coordinates of sf geometry/geometries}
\usage{
get_x(g)
}
\arguments{
\item{g}{an sf geometry}
}
\description{
Exported for internal reasons. Not intended for general use.
}
