% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{get_y}
\alias{get_y}
\title{Obtain y coordinates of sf geometry/geometries}
\usage{
get_y(g)
}
\arguments{
\item{g}{an sf geometry}
}
\description{
Exported for internal reasons. Not intended for general use.
}
