\name{getLegend}
\alias{getLegend}
\title{Function to get legend information for each scale}
\usage{
getLegend(mb)
}
\arguments{
  \item{mb}{single entry from ggplot2:::guides_merge() list
  of legend data}
}
\value{
list of legend information, NULL if guide=FALSE.
}
\description{
Function to get legend information for each scale
}

