% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom2trace.R
\name{group2NA}
\alias{group2NA}
\title{Utility functions
--------------------------------------------------------------------------}
\usage{
group2NA(data)
}
\description{
Drawing ggplot2 geoms with a group aesthetic is most efficient in
plotly when we convert groups of things that look the same to
vectors with NA.
}

