\name{group2trace}
\alias{group2trace}
\title{Convert 1 ggplot2 group to 1 plotly trace.}
\usage{
group2trace(df, params, geom)
}
\arguments{
  \item{df}{data.frame.}

  \item{params}{list of defaults.}

  \item{geom}{length 1 character.}
}
\value{
a list to be passed to plotly().
}
\description{
Convert 1 ggplot2 group to 1 plotly trace.
}
\author{
Toby Dylan Hocking
}

