% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trace.R
\name{heatmap}
\alias{heatmap}
\title{Create a 'heatmap' trace object}
\usage{
heatmap(z = NULL, x = NULL, y = NULL, name = NULL, zauto = NULL,
  zmin = NULL, zmax = NULL, colorscale = NULL, reversescale = NULL,
  showscale = NULL, colorbar = NULL, zsmooth = NULL, opacity = NULL,
  connectgaps = NULL, xaxis = NULL, yaxis = NULL, showlegend = NULL,
  stream = NULL, text = NULL, visible = NULL, x0 = NULL, dx = NULL,
  y0 = NULL, dy = NULL, xtype = NULL, ytype = NULL)
}
\arguments{
\item{z,}{x, y, name, zauto, zmin, zmax, colorscale, reversescale, showscale,
colorbar, zsmooth, opacity, connectgaps, xaxis, yaxis, showlegend, stream,
text, visible, x0, dx, y0, dy, xtype, ytype}
}
\description{
Create a 'heatmap' trace object
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#heatmap}
}

