% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hide_guides}
\alias{hide_guides}
\title{Hide guides (legends and colorbars)}
\usage{
hide_guides(p)
}
\arguments{
\item{p}{a plotly object.}
}
\description{
Hide guides (legends and colorbars)
}
\seealso{
\code{\link[=hide_legend]{hide_legend()}}, \code{\link[=hide_colorbar]{hide_colorbar()}}
}
