% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hide_legend}
\alias{hide_legend}
\title{Hide legend}
\usage{
hide_legend(p)
}
\arguments{
\item{p}{a plotly object.}
}
\description{
Hide legend
}
\examples{
\dontshow{if (interactive() || !identical(.Platform$OS.type, "windows")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

p <- plot_ly(mtcars, x = ~wt, y = ~cyl, color = ~factor(cyl))
hide_legend(p)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=hide_colorbar]{hide_colorbar()}}
}
