% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_data.R
\name{highlight_key}
\alias{highlight_key}
\title{Highlight/query data based on primary key}
\usage{
highlight_key(x, ...)
}
\arguments{
\item{x}{a plotly visualization or a \code{data.frame}.}

\item{...}{arguments passed to
\code{\link[crosstalk:SharedData]{crosstalk::SharedData$new()}}.}
}
\value{
An object of class \code{\link[crosstalk:SharedData]{crosstalk::SharedData}}.
}
\description{
This function simply creates an object of class
\code{\link[crosstalk:SharedData]{crosstalk::SharedData}}. The reason it exists is to
make it easier to teach others how to leverage its functionality in plotly.
It also makes it more discoverable if one is already aware of \code{\link[=highlight]{highlight()}}.
}
\seealso{
\link{highlight}
}
\author{
Carson Sievert
}
