% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trace.R
\name{histogram}
\alias{histogram}
\title{Create a 'histogram' trace object}
\usage{
histogram(x = NULL, y = NULL, histnorm = NULL, histfunc = NULL,
  name = NULL, orientation = NULL, autobinx = NULL, nbinsx = NULL,
  xbins = NULL, autobiny = NULL, nbinsy = NULL, ybins = NULL,
  text = NULL, error_y = NULL, error_x = NULL, marker = NULL,
  opacity = NULL, xaxis = NULL, yaxis = NULL, showlegend = NULL,
  stream = NULL, visible = NULL, xsrc = NULL, ysrc = NULL)
}
\arguments{
\item{x,}{y, histnorm, histfunc, name, orientation, autobinx, nbinsx, xbins,
autobiny, nbinsy, ybins, text, error_y, error_x, marker, opacity, xaxis,
yaxis, showlegend, stream, visible, xsrc, ysrc}
}
\description{
Create a 'histogram' trace object
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#histogram}
}

