% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trace.R
\name{histogram2d}
\alias{histogram2d}
\title{Create a 'histogram2d' trace object}
\usage{
histogram2d(x = NULL, y = NULL, histnorm = NULL, histfunc = NULL,
  name = NULL, autobinx = NULL, nbinsx = NULL, xbins = NULL,
  autobiny = NULL, nbinsy = NULL, ybins = NULL, colorscale = NULL,
  reversescale = NULL, showscale = NULL, colorbar = NULL, zauto = NULL,
  zmin = NULL, zmax = NULL, zsmooth = NULL, opacity = NULL,
  xaxis = NULL, yaxis = NULL, showlegend = NULL, stream = NULL,
  visible = NULL, xsrc = NULL, ysrc = NULL)
}
\arguments{
\item{x,}{y, histnorm, histfunc, name, autobinx, nbinsx, xbins, autobiny,
nbinsy, ybins, colorscale, reversescale, showscale, colorbar, zauto, zmin,
zmax, zsmooth, opacity, xaxis, yaxis, showlegend, stream, visible, xsrc, ysrc}
}
\description{
Create a 'histogram2d' trace object
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#histogram2d}
}

