% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{knit_print.plotly_built}
\alias{knit_print.plotly_built}
\title{Print a 'built' plotly object in a knitr doc}
\usage{
knit_print.plotly_built(x, options, ...)
}
\arguments{
\item{x}{a plotly object}

\item{...}{additional arguments}
}
\description{
Print a 'built' plotly object in a knitr doc
}

