% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{knit_print.plotly_grid}
\alias{knit_print.plotly_grid}
\title{Embed a plotly grid as an iframe in a knitr doc}
\usage{
knit_print.plotly_grid(x, options, ...)
}
\arguments{
\item{x}{a plotly grid object}

\item{options}{knitr options.}

\item{...}{placeholder.}
}
\description{
Embed a plotly grid as an iframe in a knitr doc
}
\references{
https://github.com/yihui/knitr/blob/master/vignettes/knit_print.Rmd
}
