% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{knit_print.plotly_hash}
\alias{knit_print.plotly_hash}
\title{Print a plotly object in a knitr doc}
\usage{
knit_print.plotly_hash(x, options, ...)
}
\arguments{
\item{x}{a plotly object}

\item{options}{knitr options.}

\item{...}{additional arguments}
}
\description{
Print a plotly object in a knitr doc
}

