% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotly.R
\name{knit_print.plotly_response}
\alias{knit_print.plotly_response}
\title{Embed a plotly iframe into an R markdown document via \code{knit_print}}
\usage{
knit_print.plotly_response(x, options, ...)
}
\arguments{
\item{x}{named list of ggplots and option lists to pass to \code{animint2dir}.}

\item{options}{knitr options.}

\item{...}{placeholder.}
}
\description{
Embed a plotly iframe into an R markdown document via \code{knit_print}
}
\references{
https://github.com/yihui/knitr/blob/master/vignettes/knit_print.Rmd
}

