% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{last_plot}
\alias{last_plot}
\title{Retrive and create the last plotly (or ggplot).}
\usage{
last_plot(data = NULL)
}
\arguments{
\item{data}{(optional) a data frame with a class of plotly (and a plotly_hash attribute).}
}
\description{
This function was deprecated in version 4.0.
}

