% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/layout.R
\name{layout}
\alias{layout}
\title{Create a layout object.}
\usage{
layout(title = NULL, titlefont = NULL, font = NULL, showlegend = NULL,
  autosize = NULL, width = NULL, height = NULL, xaxis = NULL,
  yaxis = NULL, legend = NULL, annotations = NULL, margin = NULL,
  paper_bgcolor = NULL, plot_bgcolor = NULL, hovermode = NULL,
  dragmode = NULL, separators = NULL, barmode = NULL, bargap = NULL,
  bargroupgap = NULL, barnorm = NULL, boxmode = NULL, boxgap = NULL,
  boxgroupgap = NULL, radialaxis = NULL, angularaxis = NULL,
  scene = NULL, direction = NULL, orientation = NULL,
  hidesources = NULL)
}
\arguments{
\item{title,}{titlefont, font, showlegend, autosize, width, height, xaxis,
yaxis, legend, annotations, margin, paper_bgcolor, plot_bgcolor, hovermode,
dragmode, separators, barmode, bargap, bargroupgap, barnorm, boxmode, boxgap,
boxgroupgap, radialaxis, angularaxis, scene, direction, orientation, hidesources}
}
\description{
A layout object by itself is not sufficient for creating a plot.
A layout object must be added to a plotly trace object.
}
\examples{
# TODO: how do we ensure a single escape character goes in response body?
histogram(rnorm(100)) + layout(title = "$X \\\\sim N(0, 1)$")
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#layout}
}

