% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/layout.R
\name{legend}
\alias{legend}
\title{Create a legend object.}
\usage{
legend(x = NULL, y = NULL, traceorder = NULL, font = NULL,
  bgcolor = NULL, bordercolor = NULL, borderwidth = NULL, xref = NULL,
  yref = NULL, xanchor = NULL, yanchor = NULL)
}
\arguments{
\item{x,}{y, traceorder, font, bgcolor, bordercolor, borderwidth, xref, yref,
xanchor, yanchor}
}
\description{
Available in \link{layout}.
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#legend}
}

