% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aux.R
\name{line}
\alias{line}
\title{Create a 'line' auxiliary object}
\usage{
line(color = NULL, width = NULL, dash = NULL, opacity = NULL,
  shape = NULL, smoothing = NULL, outliercolor = NULL,
  outlierwidth = NULL)
}
\arguments{
\item{color,}{width, dash, opacity, shape, smoothing, outliercolor, outlierwidth}
}
\description{
Available in \link{scatter}, \link{box}, \link{contour},
\link{histogram2dcontour}, \link{scatter3d}, \link{marker}.
}
\examples{
\dontrun{

}
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#line}
}

