\docType{data}
\name{lty2dash}
\alias{lty2dash}
\title{Convert R lty line type codes to plotly "dash" codes.}
\format{\preformatted{ Named chr [1:25] "solid" "dash" "dot" "dashdot" "longdash" ...
 - attr(*, "names")= chr [1:25] "1" "2" "3" "4" ...
}}
\usage{
lty2dash
}
\description{
Convert R lty line type codes to plotly "dash" codes.
}
\keyword{datasets}

