% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make_ggplot_build2}
\alias{make_ggplot_build2}
\title{ggplot build function with enhanced return}
\usage{
make_ggplot_build2(plot)
}
\arguments{
\item{plot}{ggplot2 plot}
}
\value{
List with (data, panel, plot, prestats.data) where prestats.data
is the data as it is prior to calculate_stats() call
}
\description{
This function builds on top of ggplot2::ggplot_build by
Hadley Wickham <h.wickham@gmail.com> and Winston Chang <winston@stdout.org>
(http://ggplot2.org, https://github.com/hadley/ggplot2).
}
\keyword{internal}

