% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/layout.R
\name{margin}
\alias{margin}
\title{Create a margin object.}
\usage{
margin(l = NULL, r = NULL, b = NULL, t = NULL, pad = NULL,
  autoexpand = NULL, data = NULL, layout = NULL)
}
\arguments{
\item{l,}{r, b, t, pad, autoexpand, data, layout}
}
\description{
Available in \link{layout}.
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#margin}
}

