% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aux.R
\name{marker}
\alias{marker}
\title{Create a 'marker' auxiliary object}
\usage{
marker(color = NULL, size = NULL, symbol = NULL, line = NULL,
  opacity = NULL, sizeref = NULL, sizemode = NULL, colorscale = NULL,
  cauto = NULL, cmin = NULL, cmax = NULL, outliercolor = NULL,
  maxdisplayed = NULL)
}
\arguments{
\item{color,}{size, symbol, line, opacity, sizeref, sizemode, colorscale,
cauto, cmin, cmax, outliercolor, maxdisplayed}
}
\description{
Available in \link{scatter}, \link{bar}, \link{histogram}, \link{box},
\link{area}, \link{scatter3d}.
}
\examples{
\dontrun{

}
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#marker}
}

