\name{merge_recurse}
\alias{merge_recurse}
\title{Function to merge a list of data frames (from the reshape package)}
\usage{
merge_recurse(dfs, ...)
}
\arguments{
  \item{dfs}{list of data frames}

  \item{...}{other arguments to merge}
}
\value{
data frame of merged lists
}
\description{
Function to merge a list of data frames (from the reshape
package)
}

