% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/marker_conversion.R
\name{paramORdefault}
\alias{paramORdefault}
\title{Convert ggplot params to plotly.}
\usage{
paramORdefault(params, aesVec, defaults)
}
\arguments{
\item{params}{named list ggplot names -> values.}

\item{aesVec}{vector mapping ggplot names to plotly names.}

\item{defaults}{named list ggplot names -> values.}
}
\value{
named list.
}
\description{
Convert ggplot params to plotly.
}
\author{
Toby Dylan Hocking
}

