% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{plotly-data}
\alias{arrange_.plotly}
\alias{distinct_.plotly}
\alias{filter_.plotly}
\alias{group_by_.plotly}
\alias{groups.plotly}
\alias{mutate_.plotly}
\alias{plotly-data}
\alias{rename_.plotly}
\alias{select_.plotly}
\alias{slice_.plotly}
\alias{summarise_.plotly}
\alias{transmute_.plotly}
\alias{ungroup.plotly}
\title{Data manipulation verbs for plotly.}
\usage{
groups.plotly(x)

group_by_.plotly(.data, ..., .dots, add = FALSE)

ungroup.plotly(x)

summarise_.plotly(.data, ..., .dots)

mutate_.plotly(.data, ..., .dots)

arrange_.plotly(.data, ..., .dots)

select_.plotly(.data, ..., .dots)

filter_.plotly(.data, ..., .dots)

distinct_.plotly(.data, ..., .dots)

slice_.plotly(.data, ..., .dots)

rename_.plotly(.data, ..., .dots)

transmute_.plotly(.data, ..., .dots)
}
\description{
Data manipulation verbs for plotly.
}
\examples{

# loading dplyr is super important!
library(dplyr)

mtcars \%>\%
  plot_ly(x = ~wt, y = ~mpg, name = "scatter trace") \%>\%
  filter(cyl == 4) \%>\%
  add_markers(x = ~wt, y = ~mpg, name = "filtered points", 
    marker = list(color = "red"))
    
library(tidyr)
economics \%>\%
  gather(variable, value, -date) \%>\%
  group_by(variable) \%>\%
  plot_ly(x = ~date, y = ~value)
  
  


}
\keyword{internal}

