% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{plotly-ggvis}
\alias{plotly-ggvis}
\title{Dplyr verbs for plotly.}
\description{
Dplyr verbs for plotly.
}
\examples{

library(dplyr)
plot_ly(mtcars, x = ~wt, y = ~mpg) \%>\%
  filter(cyl == 4) \%>\%
  add_points()

}
\keyword{internal}

