% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{plotly-package}
\alias{plotly}
\alias{plotly-package}
\title{plotly: Create Interactive Web Graphics via 'plotly.js'}
\description{
Easily translate 'ggplot2' graphs to an interactive web-based version and/or create custom web-based visualizations directly from R. Once uploaded to a 'plotly' account, 'plotly' graphs (and the data behind them) can be viewed and modified in a web browser.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://plot.ly/r}
  \item \url{https://cpsievert.github.io/plotly_book/}
  \item \url{https://github.com/ropensci/plotly}
  \item Report bugs at \url{https://github.com/ropensci/plotly/issues}
}

}
\author{
\strong{Maintainer}: Carson Sievert \email{cpsievert1@gmail.com}

Authors:
\itemize{
  \item Chris Parmer \email{chris@plot.ly} [copyright holder]
  \item Toby Hocking \email{tdhock5@gmail.com}
  \item Scott Chamberlain \email{myrmecocystus@gmail.com}
  \item Karthik Ram \email{karthik.ram@gmail.com}
  \item Marianne Corvellec \email{marianne@plot.ly}
  \item Pedro Despouy \email{pedro@plot.ly}
}

}
\keyword{internal}
