% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plotly_empty}
\alias{plotly_empty}
\title{Create a complete empty plotly graph.}
\usage{
plotly_empty(...)
}
\arguments{
\item{...}{arguments passed onto \code{\link[=plot_ly]{plot_ly()}}}
}
\description{
Useful when used with \code{\link[=subplot]{subplot()}}
}
