% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_example.R
\name{plotly_example}
\alias{plotly_example}
\title{Run a plotly example(s)}
\usage{
plotly_example(type = c("demo", "shiny", "rmd"), name, ...)
}
\arguments{
\item{type}{the type of example}

\item{name}{the name of the example (valid names depend on \code{type}).}

\item{...}{arguments passed onto the suitable method.}
}
\description{
Provides a unified interface for running demos, shiny apps, and Rmd documents
which are bundled with the package.
}
\author{
Carson Sievert
}
