% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotly_IMAGE.R
\name{plotly_IMAGE}
\alias{plotly_IMAGE}
\title{Create/Modify plotly images}
\usage{
plotly_IMAGE(x, width = 1000, height = 500, format = "png", scale = 4,
  encoded = FALSE)
}
\arguments{
\item{x}{either a plotly object or a list.}

\item{width}{Image width in pixels}

\item{height}{Image height in pixels}

\item{format}{The desired image format 'png', 'jpeg', 'svg', 'pdf', 'eps', or 'webp'}

\item{scale}{Both png and jpeg formats will be scaled beyond the specified width and height by this number.}

\item{encoded}{A boolean flag for base64 encoding of bytes}
}
\description{
The images endpoint turn a plot (which may be given in multiple forms)
into an image of the desired format.
}
\examples{
plotly_IMAGE(qplot(1:10))
}

