% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{plotly_json}
\alias{plotly_json}
\title{Inspect JSON sent to plotly.js}
\usage{
plotly_json(p = plot_ly(), jsonedit = interactive(), ...)
}
\arguments{
\item{p}{a plotly or ggplot object.}

\item{jsonedit}{use \code{listviewer::jsonedit} to view the JSON?}

\item{...}{other options passed onto \code{listviewer::jsonedit}}
}
\description{
This function is useful for obtaining/viewing/debugging JSON 
sent to plotly.js.
}
\examples{
  
plotly_json(plot_ly())
plotly_json(plot_ly(), FALSE)
}

