% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plotly_spec}
\alias{plotly_spec}
\title{Obtain JSON sent to plotly.js}
\usage{
plotly_spec(p = plot_ly(), as = c("plain", "json"), ...)
}
\arguments{
\item{p}{a plotly or ggplot object.}

\item{as}{in what format should the spec be printed?}

\item{...}{arguments passed onto \code{jsonlite::toJSON}}
}
\description{
This function returns the JSON sent to plotly.js which can be useful for 
debugging.
}
\examples{
  
plotly_spec(plot_ly(), as = "json", pretty = TRUE)
}

