% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shiny.R
\name{plotlyOutput}
\alias{plotlyOutput}
\title{Shiny ui output function}
\usage{
plotlyOutput(outputId, width = "100\%", height = "550px", offline = TRUE)
}
\arguments{
\item{outputId}{output variable to read the plot from}

\item{width}{width of the output.}

\item{height}{height of the output.}

\item{offline}{Use plotly offline, if available?}
}
\description{
Shiny ui output function
}
\seealso{
http://shiny.rstudio.com/articles/building-outputs.html
}

