% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotly.R
\name{+.plotly}
\alias{+.plotly}
\title{Method for adding together plotly objects}
\usage{
\method{+}{plotly}(x, y)
}
\arguments{
\item{x}{a plotly object}

\item{y}{another object}
}
\description{
Method for adding together plotly objects
}
\author{
Carson Sievert
}

