% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggridges.R
\name{prepare_ridge_chart}
\alias{prepare_ridge_chart}
\title{Prepare plotting data for ggridges}
\usage{
prepare_ridge_chart(
  data,
  prestats_data,
  layout,
  params,
  p,
  closed = FALSE,
  ...
)
}
\arguments{
\item{closed}{boolean, should the polygon be closed at bottom (TRUE for
geom_density_ridges2, FALSE for geom_density_ridges)}
}
\description{
Prepare plotting data for ggridges
}
