% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.api}
\alias{print.api}
\title{Print method for a 'generic' API response}
\usage{
\method{print}{api}(x, ...)
}
\arguments{
\item{x}{a list.}

\item{...}{additional arguments (currently ignored)}
}
\description{
Print method for a 'generic' API response
}
