% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.api_grid_local}
\alias{print.api_grid_local}
\title{Print a plotly grid object}
\usage{
\method{print}{api_grid_local}(x, ...)
}
\arguments{
\item{x}{a plotly grid object}

\item{...}{additional arguments (currently ignored)}
}
\description{
Print a plotly grid object
}
