% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.api_plot}
\alias{print.api_plot}
\title{Print a plot on plotly's platform}
\usage{
\method{print}{api_plot}(x, ...)
}
\arguments{
\item{x}{a plotly figure object}

\item{...}{additional arguments (currently ignored)}
}
\description{
Print a plot on plotly's platform
}

