% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaleido.R
\name{print.kaleidoScope}
\alias{print.kaleidoScope}
\title{Print method for kaleido}
\usage{
\method{print}{kaleidoScope}(x, ...)
}
\arguments{
\item{x}{a \code{\link[=kaleido]{kaleido()}} object.}

\item{...}{currently unused.}
}
\description{
S3 method for \code{\link[=kaleido]{kaleido()}}.
}
\keyword{internal}
