% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{print.offline}
\alias{print.offline}
\title{Print an "offline" (local) plotly object}
\usage{
\method{print}{offline}(x, ...)
}
\arguments{
\item{x}{an object with class 'offline'}

\item{...}{other arguments}
}
\description{
Print an "offline" (local) plotly object
}

