% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.plotly_figure}
\alias{print.plotly_figure}
\title{Print a plotly figure object}
\usage{
\method{print}{plotly_figure}(x, ...)
}
\arguments{
\item{x}{a plotly figure object}

\item{...}{additional arguments (currently ignored)}
}
\description{
Print a plotly figure object
}
