% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/axis.R
\name{radialaxis}
\alias{radialaxis}
\title{Create an radialaxis object.}
\usage{
radialaxis(range = NULL, domain = NULL, orientation = NULL,
  showline = NULL, showticklabels = NULL, tickorientation = NULL,
  ticklen = NULL, tickcolor = NULL, ticksuffix = NULL,
  endpadding = NULL, visible = NULL)
}
\arguments{
\item{range,}{domain, orientation, showline, showticklabels, tickorientation,
ticklen, tickcolor, ticksuffix, endpadding, visible}
}
\description{
Available in \link{layout}
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#radialaxis}
}

