% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R, R/pipe.R
\docType{import}
\name{mutate}
\alias{mutate}
\alias{mutate_}
\alias{transmute}
\alias{transmute_}
\alias{select}
\alias{select_}
\alias{rename}
\alias{rename_}
\alias{group_by}
\alias{group_by_}
\alias{groups}
\alias{ungroup}
\alias{summarise}
\alias{summarise_}
\alias{do}
\alias{do_}
\alias{arrange}
\alias{arrange_}
\alias{distinct}
\alias{distinct_}
\alias{slice}
\alias{slice_}
\alias{filter}
\alias{filter_}
\alias{reexports}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{arrange}}, \code{\link[dplyr:se-deprecated]{arrange_}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr:se-deprecated]{distinct_}}, \code{\link[dplyr]{do}}, \code{\link[dplyr:se-deprecated]{do_}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr:se-deprecated]{filter_}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr:se-deprecated]{group_by_}}, \code{\link[dplyr:group_data]{groups}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr:se-deprecated]{mutate_}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr:se-deprecated]{rename_}}, \code{\link[dplyr]{select}}, \code{\link[dplyr:se-deprecated]{select_}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr:se-deprecated]{slice_}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr:se-deprecated]{summarise_}}, \code{\link[dplyr]{transmute}}, \code{\link[dplyr:se-deprecated]{transmute_}}, \code{\link[dplyr:group_by]{ungroup}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}
}}

