% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res_mn}
\alias{res_mn}
\title{Minnesotan Indian Reservation Lands}
\format{
An sf data frame with 13 features and 5 fields
}
\usage{
res_mn
}
\description{
Minnesotan Indian Reservation Lands
}
\references{
\url{https://www.dot.state.mn.us/maps/gdma/gis-data.html}
}
\keyword{datasets}
