% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trace.R
\name{scatter}
\alias{scatter}
\title{Create a 'scatter' trace object}
\usage{
scatter(x = NULL, y = NULL, r = NULL, t = NULL, mode = "markers",
  name = NULL, text = NULL, error_y = NULL, error_x = NULL,
  marker = NULL, line = NULL, textposition = NULL, textfont = NULL,
  connectgaps = NULL, fill = NULL, fillcolor = NULL, opacity = NULL,
  xaxis = NULL, yaxis = NULL, showlegend = NULL, stream = NULL,
  visible = NULL, xsrc = NULL, ysrc = NULL)
}
\arguments{
\item{x,}{y, r, t, mode, name, text, error_y, error_x, marker, line,
textposition, textfont, connectgaps, fill, fillcolor, opacity, xaxis, yaxis,
showlegend, stream, visible, xsrc, ysrc}
}
\description{
Create a 'scatter' trace object
}
\examples{
\dontrun{
# by default, scatter() creates a scatterplot
(p <- with(economics, scatter(date, uempmed)))

m <- loess(uempmed ~ as.numeric(date), economics)
# change the mode to get a line plot (and use `+` to add it to the scatterplot)
p + scatter(economics$date, fitted(m), mode = "lines") +
  layout(showLegend = FALSE)
}
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#scatter}
}

