% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trace.R
\name{scatter3d}
\alias{scatter3d}
\title{Create a 'scatter3d' trace object}
\usage{
scatter3d(x = NULL, y = NULL, z = NULL, mode = NULL, name = NULL,
  text = NULL, error_z = NULL, error_y = NULL, error_x = NULL,
  marker = NULL, line = NULL, textposition = NULL, scene = NULL,
  stream = NULL, visible = NULL)
}
\arguments{
\item{x,}{y, z, mode, name, text, error_z, error_y, error_x, marker, line,
textposition, scene, stream, visible}
}
\description{
Create a 'scatter3d' trace object
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#scatter3d}
}

