% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/axis.R
\name{scene}
\alias{scene}
\title{Create an scene object.}
\usage{
scene(xaxis = NULL, yaxis = NULL, zaxis = NULL, cameraposition = NULL,
  domain = NULL, bgcolor = NULL)
}
\arguments{
\item{xaxis,}{yaxis, zaxis, cameraposition, domain, bgcolor}
}
\description{
Available in \link{layout}
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#scene}
}

