% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R, R/dev.R
\name{schema}
\alias{schema}
\title{Display plot schema}
\usage{
schema()

schema()
}
\description{
Display plot schema

The schema contains valid attributes names, their value type, 
default values (if any), and min/max values (if applicable).
}
\examples{
  
schema()
schema()
}

