% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectize.R
\name{selectize}
\alias{selectize}
\title{Add a selectize input box for highlighting selections}
\usage{
selectize(p, multiple = TRUE)
}
\description{
Add a selectize input box for highlighting selections
}
\examples{

sd <- SharedData$new(txhousing, ~city)
p <- plot_ly(sd, x = ~date, y = ~median) \%>\%
  add_lines(text = ~city, hoverinfo = "text") \%>\%
  highlight(on = "plotly_hover")

selectize(p)


# or you can use crosstalk::filter_select(), which is a special case of highlight() 
s <- crosstalk::filter_select("myId", "Pick a city", sd, ~city)
browsable(
  tagList(p, s)
)

}
\author{
Carson Sievert
}

