% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tools.R
\name{set_config_file}
\alias{set_config_file}
\title{Set keyword-value pairs in Plotly config file}
\usage{
set_config_file(plotly_domain = "", plotly_streaming_domain = "")
}
\arguments{
\item{plotly_domain}{plotly domain}

\item{plotly_streaming_domain}{plotly streaming domain}
}
\value{
List of keyword-value pairs (config)
}
\description{
Set keyword-value pairs in Plotly config file
}
\examples{
\dontrun{
set_config_file("https://kitty.plot.ly", "stream.kitty.plot.ly")
}
}

