% Generated by roxygen2 (4.0.2): do not edit by hand
\name{set_credentials_file}
\alias{set_credentials_file}
\title{Set the keyword-value pairs in Plotly credentials file}
\usage{
set_credentials_file(username = "", api_key = "", stream_ids = list("",
  ""))
}
\arguments{
\item{username}{plotly username}

\item{api_key}{plotly API key}

\item{stream_ids}{stream ids}
}
\value{
List of keyword-value pairs (credentials)
}
\description{
Set the keyword-value pairs in Plotly credentials file
}
\examples{
\dontrun{
set_credentials_file("username", "api_key", list("foo", "bar))
}
}

