% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_data.R
\name{shared_data}
\alias{shared_data}
\title{Shared data}
\usage{
shared_data(...)
}
\arguments{
\item{...}{arguments passed to crosstalk::SharedData$new()}
}
\value{
An object of class \link[crosstalk:SharedData]{crosstalk::SharedData}
}
\description{
This is simply a wrapper around \code{crosstalk::SharedData$new()} to make it easier
to use and explain conceptually.
}
