% Generated by roxygen2 (4.0.1): do not edit by hand
\name{show_credentials_file}
\alias{show_credentials_file}
\title{Read and print Plotly credentials file, wrapping get_credentials_file()}
\usage{
show_credentials_file(args = c())
}
\arguments{
\item{args}{character vector of credentials}
}
\value{
print plotly user credentials data
}
\description{
Read and print Plotly credentials file, wrapping get_credentials_file()
}

