% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toRGB.R
\name{showRGB}
\alias{showRGB}
\title{View colors already formatted by toRGB()}
\usage{
showRGB(x, ...)
}
\arguments{
\item{x}{character string specifying color(s).}

\item{...}{arguments passed along to \code{scales::show_col}.}
}
\description{
Useful for viewing colors after they've been converted to plotly.js'
color format -- "rgba(255, 255, 255, 1)"
}
\examples{
\dontshow{if (interactive() || !identical(.Platform$OS.type, "windows")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

showRGB(toRGB(colors()), labels = FALSE)
\dontshow{\}) # examplesIf}
}
\author{
Carson Sievert
}
