% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotly.R
\name{signup}
\alias{signup}
\title{Create a new plotly account.}
\usage{
signup(username, email, save = TRUE)
}
\arguments{
\item{username}{Desired username.}

\item{email}{Desired email.}

\item{save}{If request is successful, should the username & API key be
automatically stored as an environment variable in a .Rprofile?}
}
\value{
\itemize{
 \item api_key key to use with the api
 \item tmp_pw temporary password to access your plotly account
}
}
\description{
A sign up interface to plotly through the R Console.
}
\references{
https://plot.ly/rest/
}

