% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subplots.R
\name{subplot}
\alias{subplot}
\title{plotly subplots}
\usage{
subplot(x)
}
\arguments{
\item{x}{a list of plotly (and optionally layout) objects}
}
\description{
plotly subplots
}
\examples{
\dontrun{
p1 <- ggplotly(qplot(rnorm(100)))
p2 <- histogram(rnorm(100), xaxis(domain = c(0, 0.5)))
ps <- list(
  plot1 = p1,
  plot2 = p2,
  layout(plot1 = domain(0, ))
)
}
}
\author{
Carson Sievert
}

