% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trace.R
\name{surface}
\alias{surface}
\title{Create a 'surface' trace object}
\usage{
surface(z = NULL, x = NULL, y = NULL, name = NULL, colorscale = NULL,
  scene = NULL, stream = NULL, visible = NULL, type = NULL)
}
\arguments{
\item{z,}{x, y, name, colorscale, scene, stream, visible, type}
}
\description{
Create a 'surface' trace object
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#surface}
}

