% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers2traces.R
\name{to_basic}
\alias{to_basic}
\title{Convert a geom, which are special cases of other geoms, to "basic" geoms.
--------------------------------------------------------------------------}
\usage{
to_basic(data, prestats_data, layout, params, ...)
}
\description{
Convert a geom, which are special cases of other geoms, to "basic" geoms.
--------------------------------------------------------------------------
}

