% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_conversion.R
\name{toFill}
\alias{toFill}
\title{Use default ggplot colour for fill (gray20) if not declared}
\usage{
toFill(x, alpha = 1)
}
\arguments{
\item{x}{character for colour}

\item{alpha}{transparency alpha}
}
\value{
hexadecimal colour value
}
\description{
Use default ggplot colour for fill (gray20) if not declared
}

