% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{toWebGL}
\alias{toWebGL}
\title{Convert trace types to WebGL}
\usage{
toWebGL(p, warn = TRUE)
}
\arguments{
\item{p}{a plotly or ggplot object.}

\item{warn}{should a warning be produced if the trace type doesn't have 
a WebGL equivalent?}
}
\description{
Convert trace types to WebGL
}
\examples{

# currently no bargl trace type
toWebGL(qplot(1:10))
toWebGL(qplot(1:10, 1:10))

}

