% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{toWidget}
\alias{toWidget}
\title{Convert a plotly object to an htmlwidget object}
\usage{
toWidget(x)
}
\arguments{
\item{x}{a plotly object.}
}
\description{
Users shouldn't need to use this function. It's exported for internal reasons.
}

