% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/axis.R
\name{xaxis}
\alias{xaxis}
\title{Create an xaxis object.}
\usage{
xaxis(title = NULL, titlefont = NULL, range = NULL, domain = NULL,
  type = NULL, rangemode = NULL, autorange = NULL, showgrid = NULL,
  zeroline = NULL, showline = NULL, autotick = NULL, nticks = NULL,
  ticks = NULL, showticklabels = NULL, tick0 = NULL, dtick = NULL,
  ticklen = NULL, tickwidth = NULL, tickcolor = NULL, tickangle = NULL,
  tickfont = NULL, exponentformat = NULL, showexponent = NULL,
  mirror = NULL, gridcolor = NULL, gridwidth = NULL,
  zerolinecolor = NULL, zerolinewidth = NULL, linecolor = NULL,
  linewidth = NULL, anchor = NULL, overlaying = NULL, side = NULL,
  position = NULL, showbackground = NULL, backgroundcolor = NULL,
  showspikes = NULL, spikesides = NULL, spikethickness = NULL)
}
\arguments{
\item{title,}{titlefont, range, domain, type, rangemode, autorange, showgrid,
zeroline, showline, autotick, nticks, ticks, showticklabels, tick0, dtick,
ticklen, tickwidth, tickcolor, tickangle, tickfont, exponentformat,
showexponent, mirror, gridcolor, gridwidth, zerolinecolor, zerolinewidth,
linecolor, linewidth, anchor, overlaying, side, position, showbackground,
backgroundcolor, showspikes, spikesides, spikethickness}
}
\description{
Available in \link{scene} and \link{layout}
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#xaxis}
}

