% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aux.R
\name{ybins}
\alias{ybins}
\title{Create a 'ybins' auxiliary object}
\usage{
ybins(start = NULL, end = NULL, size = NULL)
}
\arguments{
\item{start,}{end, size}
}
\description{
Available in \link{histogram}, \link{histogram2d}, \link{histogram2dcontour}
}
\examples{
\dontrun{

}
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#ybins}
}

