% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_annotations}
\alias{add_annotations}
\title{Add an annotation(s) to a plot}
\usage{
add_annotations(p, text = NULL, ..., data = NULL, inherit = TRUE)
}
\arguments{
\item{p}{a plotly object}

\item{text}{annotation text (required).}

\item{...}{these arguments are documented at
\url{https://github.com/plotly/plotly.js/blob/master/src/components/annotations/attributes.js}}

\item{data}{a data frame.}

\item{inherit}{inherit attributes from \code{\link[=plot_ly]{plot_ly()}}?}
}
\description{
Add an annotation(s) to a plot
}
\author{
Carson Sievert
}
