% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_polygons}
\alias{add_polygons}
\title{Add polygons to a plotly vis}
\usage{
add_polygons(p, mode = "lines", ...)
}
\arguments{
\item{p}{a plotly or ggplot object.}

\item{mode}{Determines the drawing mode for this scatter trace 
\url{https://plot.ly/r/reference/#scatter-mode}}

\item{...}{These arguments are documented in the references section below.
Note that acceptable arguments depend on the trace type.}
}
\description{
Add polygons to a plotly vis
}
\examples{

ggplot2::map_data("world", "canada") \%>\%
  group_by(group) \%>\%
  plot_ly(x = ~long, y = ~lat) \%>\%
  add_polygons(hoverinfo = "none") \%>\%
  add_markers(text = ~paste(name, "<br />", pop), hoverinfo = "text",
    data = maps::canada.cities) \%>\%
  layout(showlegend = FALSE)
}

