% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_trace}
\alias{add_trace}
\title{Add trace(s) to a plotly visualization}
\usage{
add_trace(p, ..., color, symbol, size, linetype, data = NULL)
}
\arguments{
\item{p}{a plotly or ggplot object.}

\item{...}{These arguments are documented in the references section below.
Note that acceptable arguments depend on the trace type.}

\item{color}{Either a variable name or a vector to use for color mapping.}

\item{symbol}{Either a variable name or a (discrete) vector to use for symbol encoding.}

\item{size}{A variable name or numeric vector to encode the size of markers.}

\item{linetype}{Either a variable name or a (discrete) vector to use for linetype encoding.}

\item{data}{A data frame to associate with this trace (optional). If not 
provided, arguments are evaluated using the data frame in \code{\link{plot_ly}()}.}
}
\description{
Add trace(s) to a plotly visualization
}
\examples{

p <- plot_ly(economics, x = ~date, y = ~uempmed)
p
p \%>\% add_points()
p \%>\% add_lines()
p \%>\% add_text(text = ".")

# attributes declared in plot_ly() carry over to downstream traces
plot_ly(economics, x = ~date, y = ~uempmed) \%>\% 
  add_points(color = ~pop) \%>\%
  add_lines(line = list(color = "red"))
  

}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/r/reference/}
}
\seealso{
\code{\link{plot_ly}()}
}

