% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotly.R
\name{add_trace}
\alias{add_trace}
\title{Add a trace to a plotly visualization}
\usage{
add_trace(p = get_plot(strict = FALSE), ..., data = NULL)
}
\arguments{
\item{p}{A plotly visualization.}

\item{...}{Visual properties.
All arguments documented in the references section below are supported.
In addition, there are special arguments which map variables to visual
aethestics in a similar style to ggplot2 (such as \code{color}).}

\item{data}{A data frame (optional).}
}
\description{
Add a trace to a plotly visualization
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/r/reference/}
}

