% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_trace}
\alias{add_area}
\alias{add_lines}
\alias{add_markers}
\alias{add_paths}
\alias{add_polygons}
\alias{add_ribbons}
\alias{add_text}
\alias{add_trace}
\title{Add trace(s) to a plotly visualization}
\usage{
add_trace(p, ..., color, symbol, size, linetype, data = NULL)

add_markers(p, x = NULL, y = NULL, ...)

add_paths(p, x = NULL, y = NULL, ...)

add_text(p, x = NULL, y = NULL, text = NULL, ...)

add_lines(p, x = NULL, y = NULL, ...)

add_polygons(p, x = NULL, y = NULL, ...)

add_ribbons(p, x = NULL, ymin = NULL, ymax = NULL, ...)

add_area(p, x = NULL, ymax = NULL, ...)
}
\arguments{
\item{p}{a plotly or ggplot object.}

\item{...}{These arguments are documented in the references section below.
Note that acceptable arguments depend on the trace type.}

\item{color}{Either a variable name or a vector to use for color mapping.}

\item{symbol}{Either a variable name or a (discrete) vector to use for symbol encoding.}

\item{size}{A variable name or numeric vector to encode the size of markers.}

\item{linetype}{Either a variable name or a (discrete) vector to use for linetype encoding.}

\item{data}{A data frame to associate with this trace (optional). If not 
provided, arguments are evaluated using the data frame in \code{\link{plot_ly}()}.}

\item{x}{the x variable.}

\item{y}{the y variable.}

\item{text}{textual labels.}

\item{ymin}{a variable used to define the lower boundary of a polygon.}

\item{ymax}{a variable used to define the upper boundary of a polygon.}
}
\description{
Add trace(s) to a plotly visualization

Add points to a plotly vis

Add paths to a plotly vis

Add text to a plotly vis

Equivalent to \code{add_paths}, but with the x-values sorted.

Add polygons to a plotly vis

Ribbon plots

Equivalent to \link{add_ribbon} with \code{ymin} set to 0.
}
\examples{

p <- plot_ly(economics, x = ~date, y = ~uempmed)
p
p \%>\% add_markers()
p \%>\% add_lines()
p \%>\% add_text(text = ".")

# attributes declared in plot_ly() carry over to downstream traces
plot_ly(economics, x = ~date, y = ~uempmed) \%>\% 
  add_markers(color = ~pop) \%>\%
  add_lines(line = list(color = "red"))
  


ggplot2::map_data("world", "canada") \%>\%
  group_by(group) \%>\%
  plot_ly(x = ~long, y = ~lat) \%>\%
  add_polygons(hoverinfo = "none") \%>\%
  add_markers(text = ~paste(name, "<br />", pop), hoverinfo = "text",
    data = maps::canada.cities) \%>\%
  layout(showlegend = FALSE)
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/r/reference/}
}
\seealso{
\code{\link{plot_ly}()}
}

