% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{aes2line}
\alias{aes2line}
\title{Convert ggplot2 aes to line parameters.}
\format{\preformatted{ Named chr [1:4] "dash" "color" "width" "shape"
 - attr(*, "names")= chr [1:4] "linetype" "colour" "size" "direction"
}}
\usage{
aes2line
}
\description{
Convert ggplot2 aes to line parameters.
}
\keyword{datasets}

