\docType{data}
\name{aes2line}
\alias{aes2line}
\title{Convert ggplot2 aes to line parameters.}
\format{\preformatted{ Named chr [1:3] "dash" "color" "width"
 - attr(*, "names")= chr [1:3] "linetype" "colour" "size"
}}
\usage{
aes2line
}
\description{
Convert ggplot2 aes to line parameters.
}
\keyword{datasets}

