% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{aes2marker}
\alias{aes2marker}
\title{Convert ggplot2 aes to plotly "marker" codes.}
\format{\preformatted{ Named chr [1:6] "opacity" "color" "size" "sizeref" "sizemode" "symbol"
 - attr(*, "names")= chr [1:6] "alpha" "colour" "size" "sizeref" ...
}}
\usage{
aes2marker
}
\description{
Convert ggplot2 aes to plotly "marker" codes.
}
\keyword{datasets}

