% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{animate}
\alias{animate}
\title{Trigger an animation}
\usage{
animate(..., options = animationOpts())
}
\arguments{
\item{...}{One of the following 
\itemize{
 \item any number of plotly/ggplot2 objects.
 \item a list of plotly/ggplot2 objects.
 \item a tibble with one list-column of plotly/ggplot2 objects.
}}

\item{options}{animation options. See \code{\link{animationOpts}()}.}
}
\description{
Trigger an animation
}
\examples{

cols <- I(c("red", "blue"))
animate(
  plot_ly(x = c(1, 2), y = c(1, 2), color = cols),
  plot_ly(x = c(2, 1), y = c(2, 1), color = cols),
  plot_ly(x = c(1, 2), y = c(1, 2), color = cols)
) 

# TODO: prepopulate layout$xaxisid$range so stuff doesn't go running off the page
mtcars \%>\%
 split(.$cyl) \%>\%
 lapply(function(d) plot_ly(d, x = ~wt, y = ~mpg)) \%>\%
 animate(options = animationOpts(redraw = T))

}
\author{
Carson Sievert
}
\seealso{
\code{\link{animationOpts}()}
}

