% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{upload_plot}
\alias{upload_plot}
\alias{download_plot}
\alias{overwrite_plot}
\alias{upload_grid}
\alias{overwrite_grid}
\alias{download_grid}
\title{Tools for working with plotly's REST API (v2)}
\usage{
upload_plot(p = last_plot(), filename = NULL, sharing = c("public",
  "private", "secret"), ...)

download_plot(id, username)

overwrite_plot()

upload_grid(x, filename = as.character(substitute(x)), sharing = c("public",
  "private", "secret"), ...)

overwrite_grid(x, filename = as.character(substitute(x)), ...)

download_grid(id, username)
}
\arguments{
\item{p}{either a ggplot object, a plotly object, or a list.}

\item{filename}{character string which names the file.}

\item{sharing}{If 'public', anyone can view this graph. It will appear in 
your profile and can appear in search engines. You do not need to be
logged in to Plotly to view this chart.
If 'private', only you can view this plot. It will not appear in the
Plotly feed, your profile, or search engines. You must be logged in to 
Plotly to view this graph. You can privately share this graph with other 
Plotly users in your online Plotly account and they will need to be logged 
in to view this plot.
If 'secret', anyone with this secret link can view this chart. It will
not appear in the Plotly feed, your profile, or search engines. 
If it is embedded inside a webpage or an IPython notebook, anybody who is 
viewing that page will be able to view the graph. 
You do not need to be logged in to view this plot.}

\item{id}{a filename id.}

\item{username}{a plotly username.}
}
\description{
Convenience functions for uploading, downloading, and overwriting 3 
kinds of important plotly objects: grids, plots, and dashboards. The actual 
REST API exposes 7 kinds of objects,
}
\details{
plotly's web interface is built entirely upon it's REST API which exposes 3
"classes" of important objects: grids, plots, and dashboards.
To learn how to use the API more directly, see 
\url{http://moderndata.plot.ly/simple-rest-apis-for-charts-and-datasets/}
}
\examples{
\dontrun{
# upload a data frame and start creating via plotly's platform
m <- upload_grid(mtcars)

# or download data from the platform into R
download_grid("14705", "cpsievert")

# upload a plot to plotly's web platforam
p <- plot_ly(mtcars, x = ~factor(vs))
r <- upload_plot(p, filename = "mtcars-bar-plot")
# note that you can always obtain about the remote figure
str(r)

# or pull a plot down from the platform and work with it locally
download_plot("14708", "cpsievert")

}

}
\references{
\url{https://api.plot.ly/v2}, \url{http://moderndata.plot.ly/simple-rest-apis-for-charts-and-datasets/}
}
\seealso{
\code{\link{signup}()}
}
\author{
Carson Sievert
}
