% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{as.widget}
\alias{as.widget}
\title{Convert a plotly object to an htmlwidget object}
\usage{
as.widget(x, ...)
}
\arguments{
\item{x}{a plotly object.}

\item{...}{other options passed onto \code{htmlwidgets::createWidget}}
}
\description{
As of version 4.0.0, plotly objects are now htmlwidget objects,
so users shouldn't have to use this function directly.
}

