% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{as.widget}
\alias{as.widget}
\title{Convert a plotly object to an htmlwidget object}
\usage{
as.widget(x, ...)
}
\arguments{
\item{x}{a plotly object.}

\item{...}{other options passed onto \code{htmlwidgets::createWidget}}
}
\description{
Convert a plotly object to an htmlwidget object
}
\examples{
\dontrun{
p <- plot_ly(mtcars, x = mpg, y = disp, mode = "markers")
htmlwidgets::saveWidget(as.widget(p), "index.html")
}

}

