% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{as.widget}
\alias{as.widget}
\title{Convert a plotly object to an htmlwidget object}
\usage{
as.widget(x, ...)
}
\arguments{
\item{x}{a plotly object.}

\item{...}{other options passed onto \code{htmlwidgets::createWidget}}
}
\description{
This function was deprecated in  4.0.0,
as plotly objects are now htmlwidget objects,
so there is no need to convert them.
}
