% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotly.R
\name{bbox}
\alias{bbox}
\title{Estimate bounding box of a rotated string}
\usage{
bbox(txt = "foo", angle = 0, size = 12)
}
\arguments{
\item{txt}{a character string of length 1}

\item{angle}{sets the angle of the tick labels with respect to the
horizontal (e.g., \code{tickangle} of -90 draws the tick labels vertically)}

\item{size}{vertical size of a character}
}
\description{
Estimate bounding box of a rotated string
}
\references{
https://www.dropbox.com/s/nc6968prgw8ne4w/bbox.pdf?dl=0
}
