% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aux.R
\name{colorbar}
\alias{colorbar}
\title{Create a 'colorbar' auxiliary object}
\usage{
colorbar(title = NULL, titleside = NULL, titlefont = NULL,
  thickness = NULL, thicknessmode = NULL, len = NULL, lenmode = NULL,
  autotick = NULL, nticks = NULL, ticks = NULL, showticklabels = NULL,
  tick0 = NULL, dtick = NULL, ticklen = NULL, tickwidth = NULL,
  tickcolor = NULL, tickangle = NULL, tickfont = NULL,
  exponentformat = NULL, showexponent = NULL, x = NULL, y = NULL,
  xanchor = NULL, yanchor = NULL, bgcolor = NULL, outlinecolor = NULL,
  outlinewidth = NULL, bordercolor = NULL, borderwidth = NULL,
  xpad = NULL, ypad = NULL)
}
\arguments{
\item{title,}{titleside, titlefont, thickness, thicknessmode, len, lenmode,
autotick, nticks, ticks, showticklabels, tick0, dtick, ticklen, tickwidth,
tickcolor, tickangle, tickfont, exponentformat, showexponent, x, y, xanchor,
yanchor, bgcolor, outlinecolor, outlinewidth, bordercolor, borderwidth,
xpad, ypad}
}
\description{
Available in \link{heatmap}, \link{contour}, \link{histogram2d}, \link{histogram2dcontour}
}
\examples{
\dontrun{

}
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#colorbar}
}

