% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{config}
\alias{config}
\title{Set the default configuration for plotly}
\usage{
config(p, ..., collaborate = TRUE, cloud = FALSE)
}
\arguments{
\item{p}{a plotly object}

\item{...}{these arguments are documented at 
\url{https://github.com/plotly/plotly.js/blob/master/src/plot_api/plot_config.js}}

\item{collaborate}{include the collaborate mode bar button (unique to the R pkg)?}

\item{cloud}{include the send data to cloud button?}
}
\description{
Set the default configuration for plotly
}
\examples{

config(plot_ly(), displaylogo = FALSE, collaborate = FALSE)

}
\author{
Carson Sievert
}
