% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{config}
\alias{config}
\title{Set the default configuration for plotly}
\usage{
config(p = last_plot(), ...)
}
\arguments{
\item{p}{a plotly object}

\item{...}{these arguments are documented at
\url{https://github.com/plotly/plotly.js/blob/master/src/plot_api/plot_config.js}}
}
\description{
Set the default configuration for plotly
}
\examples{
\dontrun{
config(plot_ly(), displaylogo = FALSE, modeBarButtonsToRemove = list('sendDataToCloud'))
}
}
\author{
Carson Sievert
}

