% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotly.R
\name{embed_notebook}
\alias{embed_notebook}
\title{Embed a plotly iframe into a IPython Notebook}
\usage{
embed_notebook(url, width = "100\%", height = "525")
}
\arguments{
\item{url}{A url pointing to a plotly graph}

\item{width}{attribute of the iframe}

\item{height}{attribute of the iframe}
}
\description{
Embed a plotly iframe into a IPython Notebook
}

