% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{event_data}
\alias{event_data}
\title{Access plotly user input event data in shiny}
\usage{
event_data(event = c("plotly_hover", "plotly_click", "plotly_selected",
  "plotly_relayout"), source = "A",
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{event}{The type of plotly event. Currently 'plotly_hover',
'plotly_click', 'plotly_selected', and 'plotly_relayout' are supported.}

\item{source}{a character string of length 1. Match the value of this string
with the source argument in \code{\link[=plot_ly]{plot_ly()}} to retrieve the
event data corresponding to a specific plot (shiny apps can have multiple plots).}

\item{session}{a shiny session object (the default should almost always be used).}
}
\description{
This function must be called within a reactive shiny context.
}
\examples{
\dontrun{
plotly_example("shiny", "event_data")
}
}
\author{
Carson Sievert
}
