% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{Export a plotly graph to a static file}
\usage{
export(p, file = "plotly.png", ...)
}
\arguments{
\item{p}{a plotly or ggplot object.}

\item{file}{a filename. See the file argument of \code{webshot::webshot} 
for valid extensions.}

\item{...}{arguments passed onto \code{webshot::webshot}}
}
\description{
Export a plotly graph to a static file
}
\examples{
\dontrun{
export(plot_ly(economics, x = ~date, y = ~pce))
}
}

