% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers2traces.R
\name{geom2trace}
\alias{geom2trace}
\title{Convert "basic" geoms to plotly.js traces.
--------------------------------------------------------------------------}
\usage{
geom2trace(data, params)
}
\description{
Convert "basic" geoms to plotly.js traces.
--------------------------------------------------------------------------
}

