% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{geom2trace}
\alias{geom2trace}
\title{Convert basic geoms to traces.}
\format{\preformatted{List of 6
 $ path     :function (data, params)  
 $ polygon  :function (data, params)  
 $ point    :function (data, params)  
 $ bar      :function (data, params)  
 $ step     :function (data, params)  
 $ histogram:function (data, params)  
}}
\usage{
geom2trace
}
\description{
Convert basic geoms to traces.
}
\keyword{datasets}

