% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/figure.R
\name{get_figure}
\alias{get_figure}
\title{Request a figure object}
\usage{
get_figure(username, id)
}
\arguments{
\item{username}{corresponding username for the figure.}

\item{id}{of the Plotly figure.}
}
\description{
Figure objects work in the same way as plotly objects, but when printed,
they overwrite the already existing plotly object
(instead of creating a new plotly).
}
\examples{
\dontrun{
 # Anyone can obtain the information for a particular plot
 fig <- get_figure("cpsievert", "559")
 # If you have proper credentials, you can modify it
 layout(fig, title = paste("Modified on ", Sys.time()))
}
}
\references{
https://plot.ly/rest/
}

