% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_figure}
\alias{get_figure}
\title{Request a figure object}
\usage{
get_figure(username, id)
}
\arguments{
\item{username}{corresponding username for the figure.}

\item{id}{of the Plotly figure.}
}
\description{
Deprecated: see \code{\link[=api_download_plot]{api_download_plot()}}.
}
