% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotly.R
\name{get_plot}
\alias{get_plot}
\title{Obtain underlying data of plotly object}
\usage{
get_plot(data = NULL, strict = TRUE)
}
\arguments{
\item{data}{a data frame with a class of plotly (and a plotly_hash attribute).}
}
\description{
Given a data frame with a class of plotly, this function returns the arguments
and/or data used to create the plotly. If no data frame is provided,
the last plotly object created in this R session is returned (if it exists).
}

