% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotly.R
\name{gg2list}
\alias{gg2list}
\title{Convert a ggplot to a list.}
\usage{
gg2list(p, width = NULL, height = NULL)
}
\arguments{
\item{p}{ggplot2 plot.}

\item{width}{Width in pixels (optional, defaults to automatic sizing).}

\item{height}{Height in pixels (optional, defaults to automatic sizing).}
}
\value{
a 'built' plotly object (list with names "data" and "layout").
}
\description{
Convert a ggplot to a list.
}

