% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotly.R
\name{gg2list}
\alias{gg2list}
\title{Convert a ggplot to a list.}
\usage{
gg2list(p, width = NULL, height = NULL, tooltip = "all",
  originalData = TRUE, source = "A", ...)
}
\arguments{
\item{p}{ggplot2 plot.}

\item{width}{Width of the plot in pixels (optional, defaults to automatic sizing).}

\item{height}{Height of the plot in pixels (optional, defaults to automatic sizing).}

\item{tooltip}{a character vector specifying which aesthetic tooltips to show in the
tooltip. The default, "all", means show all the aesthetic tooltips
(including the unofficial "text" aesthetic).}

\item{originalData}{should the "original" or "scaled" data be returned?}

\item{source}{Only relevant for \link{event_data}.}

\item{...}{currently not used}
}
\value{
a 'built' plotly object (list with names "data" and "layout").
}
\description{
Convert a ggplot to a list.
}

