% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotly.R
\name{gg2list}
\alias{gg2list}
\title{Convert a ggplot to a list.}
\usage{
gg2list(p)
}
\arguments{
\item{p}{ggplot2 plot.}
}
\value{
figure object (list with names "data" and "layout").
}
\description{
Convert a ggplot to a list.
}

