% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ggplot_build2}
\alias{ggplot_build2}
\title{wrapper for build function}
\usage{
ggplot_build2(plot)
}
\arguments{
\item{plot}{ggplot2 plot}
}
\value{
List with (data, panel, plot, prestats.data) where prestats.data
is the data as it is prior to calculate_stats() call
}
\description{
It would be inefficient to perform each step found in make_ggplot_build2()
every time we want to make use of its output.
}
\keyword{internal}

