% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplotly.R
\name{ggplotly}
\alias{ggplotly}
\title{Create plotly graphs using ggplot2 syntax}
\usage{
ggplotly(p = ggplot2::last_plot())
}
\arguments{
\item{p}{a ggplot object.}
}
\description{
See up-to-date documentation and examples at
\url{https://plot.ly/ggplot2}
}
\examples{
\dontrun{
# simple example
ggiris <- qplot(Petal.Width, Sepal.Length, data = iris, color = Species)
ggplotly(ggiris)

# maps!!
data(canada.cities, package = "maps")
viz <- ggplot(canada.cities, aes(long, lat)) +
  borders(regions = "canada", name = "borders") +
  coord_equal() +
  geom_point(aes(text = name, size = pop), colour = "red",
               alpha = 1/2, name = "cities")
 ggplotly(viz)
}
}
\author{
Carson Sievert
}
\seealso{
\link{signup}, \link{plot_ly}
}

