% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{group2NA}
\alias{group2NA}
\title{Insert missing values to create trace groupings}
\usage{
group2NA(data, groupName = "group", nested = NULL,
  retrace.first = inherits(data, "GeomPolygon"))
}
\arguments{
\item{data}{a data frame.}

\item{groupName}{the name of the grouping variable as a character string}

\item{nested}{other variables that group should be nested 
(i.e., ordered) within.}

\item{retrace.first}{should the first row of each group be appended to the 
last row? This is useful for enclosing polygons with lines.}
}
\description{
If a group of traces share the same non-positional characteristics (i.e.,
color, fill, etc), it is more efficient to draw them as a single trace 
with missing values that separate the groups (instead of multiple traces).
This is a helper function for inserting missing values into a data set
}
\examples{
elong <- tidyr::gather(economics, variable, value, -date)
plot_ly(group2NA(elong, "variable"), x = ~date, y = ~value)

}

