% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{group_by}
\alias{group_by}
\title{Divide data into groups.}
\arguments{
\item{p}{a plotly visualization.}

\item{...}{variables to group by.}

\item{add}{By default, when \code{add = FALSE}, \code{group_by} will
override existing groups. To instead add to the existing groups,
use \code{add = TRUE}}
}
\description{
Divide data into groups.
}

