% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hide_colorbar}
\alias{hide_colorbar}
\title{Hide a color bar}
\usage{
hide_colorbar(p)
}
\arguments{
\item{p}{a plotly object.}
}
\description{
Hide a color bar
}
\examples{

plot_ly(economics, x = ~date, y = ~unemploy / pop, color = ~pop) \%>\%
  add_markers() \%>\%
  hide_colorbar()
}

