% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hide_colorbar}
\alias{hide_colorbar}
\title{Hide color bar(s)}
\usage{
hide_colorbar(p)
}
\arguments{
\item{p}{a plotly object.}
}
\description{
Hide color bar(s)
}
\examples{

p <- plot_ly(mtcars, x = ~wt, y = ~cyl, color = ~cyl)
hide_colorbar(p)
  
}
\seealso{
\code{\link[=hide_legend]{hide_legend()}}
}
