% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hide_legend}
\alias{hide_legend}
\title{Hide legend}
\usage{
hide_legend(p)
}
\arguments{
\item{p}{a plotly object.}
}
\description{
Hide legend
}
\examples{

p <- plot_ly(mtcars, x = ~wt, y = ~cyl, color = ~factor(cyl))
hide_legend(p)
}
\seealso{
\code{\link[=hide_colorbar]{hide_colorbar()}}
}
