% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{highlight}
\alias{highlight}
\title{Highlight graphical elements}
\usage{
highlight(p, on = "plotly_selected", off = "plotly_relayout",
  dynamic = FALSE, persistent = FALSE, defaultValues = NULL,
  color = NULL, opacityDim = 0.2, showInLegend = FALSE)
}
\arguments{
\item{p}{a plotly visualization.}

\item{on}{turn on a selection on which event(s)? Likely candidates are
'plotly_hover', 'plotly_click', 'plotly_selected'.}

\item{off}{turn off a selection on which event(s)? Likely candidates are
'plotly_unhover', 'plotly_doubleclick', 'plotly_deselect'.}

\item{dynamic}{should UI controls for managing selection aesthetics be 
included in the output?}

\item{persistent}{should selections persist (i.e., accumulate)?}

\item{defaultValues}{a vector of values for setting a "default selection".
These values should match the key attribute.}

\item{color}{character string of color(s) to use for 
highlighting selections. See \code{\link{toRGB}()} for valid color
specifications. If \code{NULL} (the default), the color of selected marks
are not altered (only their opacity).}

\item{opacityDim}{a number between 0 and 1 used to reduce the
opacity of non-selected traces (by multiplying with the existing opacity).}

\item{showInLegend}{populate an additional legend entry for the selection?}
}
\description{
Controls the visual appearance of selections deriving from a given
selection set.
}
\examples{

d <- crosstalk::SharedData$new(txhousing, ~city)
p <- ggplot(d, aes(date, median, group = city)) +
  geom_line(geom = "line")
ggplotly(p, tooltip = "city") \%>\%
  highlight(on = "plotly_hover", color = "red")

}

