% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{last_plot}
\alias{last_plot}
\title{Retrive last plotly to be modified or created}
\usage{
last_plot(...)
}
\description{
Retrive last plotly to be modified or created
}
\seealso{
\link{plotly_build}
}

