% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_plot.R
\name{last_plot}
\alias{last_plot}
\title{Retrieve the last plot to be modified or created.}
\usage{
last_plot()
}
\description{
Retrieve the last plot to be modified or created.
}
\seealso{
\code{\link[ggplot2:last_plot]{ggplot2::last_plot()}}
}
