% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotly.R
\name{layout}
\alias{layout}
\title{Add and/or modify layout of a plotly}
\usage{
layout(data = data.frame(), ...)
}
\arguments{
\item{data}{A data frame with a class of plotly.}

\item{...}{Trace arguments. Arguments are evaluated in the environment attached to
the most recent trace. See the reference below for documentation.}
}
\description{
Add and/or modify layout of a plotly
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#layout}
}

