% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{layout}
\alias{layout}
\title{Add and/or modify layout of a plotly}
\usage{
layout(p = last_plot(), ..., data = NULL, evaluate = FALSE)
}
\arguments{
\item{p}{A plotly object.}

\item{...}{Arguments to the layout object. For documentation,
see \url{https://plot.ly/r/reference/#Layout_and_layout_style_objects}}

\item{data}{A data frame to associate with this layout (optional). If not 
provided, arguments are evaluated using the data frame in \code{\link{plot_ly}()}.}

\item{evaluate}{logical. Evaluate arguments when this function is called?}
}
\description{
Add and/or modify layout of a plotly
}
\author{
Carson Sievert
}

