% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotly.R
\name{layout}
\alias{layout}
\title{Add and/or modify layout of a plotly}
\usage{
layout(p = get_plot(), ..., data = NULL, evaluate = FALSE)
}
\arguments{
\item{p}{A plotly visualization.}

\item{...}{Visual properties.
All arguments documented in the references section below are supported.
In addition, there are special arguments which map variables to visual
aethestics in a similar style to ggplot2 (such as \code{color}).}

\item{data}{A data frame (optional).}

\item{evaluate}{logical. Evaluate arguments when this function is called?}
}
\description{
Add and/or modify layout of a plotly
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/r/reference/#Layout_and_layout_style_objects}
}

