% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/offline.R
\name{offline}
\alias{offline}
\title{Plotly Offline}
\usage{
offline(p = get_plot(), height = 400, width = "100\%", out_dir = NULL,
  open_browser = interactive(), src_dir = "~/.plotly/plotlyjs")
}
\arguments{
\item{p}{a plotly object}

\item{height}{A valid CSS unit (like "100%", "600px", "auto") or a number,
which will be coerced to a string and have "px" appended.}

\item{width}{A valid CSS unit (like "100%", "600px", "auto") or a number,
which will be coerced to a string and have "px" appended.}

\item{out_dir}{a directory to place the visualization.
If \code{NULL}, a temporary directory is used when the offlin object is printed.}

\item{open_browser}{open the visualization after creating it?}

\item{src_dir}{a directory pointing to the plotlyjs source}
}
\value{
a plotly object of class "offline"
}
\description{
Create a plotly visualization that doesn't require an external plotly server.
}
\examples{
\dontrun{
p <- plot_ly(data = iris, x = Sepal.Width, y = Sepal.Length, color = Species,
  mode = "markers")
offline(p)

# offline ggplots
gg <- qplot(data = iris, x = Sepal.Width, y = Sepal.Length, color = Species)
offline(gg)
}
}
\author{
Carson Sievert
}
\references{
\url{http://purchasing.plot.ly/}
}

