% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{offline}
\alias{offline}
\title{Plotly Offline}
\usage{
offline(p, height, width, out_dir, open_browser)
}
\arguments{
\item{p}{a plotly object}

\item{height}{A valid CSS unit. (like "100\%", "600px", "auto") or a number,
which will be coerced to a string and have "px" appended.}

\item{width}{A valid CSS unit. (like "100\%", "600px", "auto") or a number,
which will be coerced to a string and have "px" appended.}

\item{out_dir}{a directory to place the visualization.
If \code{NULL}, a temporary directory is used when the offline object is printed.}

\item{open_browser}{open the visualization after creating it?}
}
\value{
a plotly object of class "offline"
}
\description{
Deprecated in version 2.0 (offline plots are now the default)
}
\author{
Carson Sievert
}
