% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_bundles.R
\name{partial_bundle}
\alias{partial_bundle}
\title{Use a partial bundle of plotly.js}
\usage{
partial_bundle(p, type = "auto", local = TRUE, minified = TRUE)
}
\arguments{
\item{p}{a plotly object.}

\item{type}{name of the (partial) bundle. The default \code{'auto'} attempts to}

\item{local}{either "local", "cdn", or "rawgit".}

\item{minified}{whether or not to use a minified js file (non-minified file can be useful for debugging plotly.js)}

\item{version}{version number (e.g. 1.30.0). See \href{https://github.com/plotly/plotly.js/releases}{here}
for a list of valid versions and changelogs.}
}
\description{
Use a partial bundle of plotly.js
}
\examples{

library(plotly)
p1 <- plot_ly(x = 1:10, y = 1:10) \%>\% add_markers()
p2 <- partial_bundle(p1)
f1 <- tempfile(fileext = ".html")
f2 <- tempfile(fileext = ".html")

save_widget <- function(p, f) {
  owd <- setwd(dirname(f))
  on.exit(setwd(owd))
  htmlwidgets::saveWidget(p, f)
  mb <- round(file.info(f)$size / 1e6, 3)
  message("File is: ", mb," MB")
}
save_widget(p1, f1)
save_widget(p2, f2)

}
\references{
\url{https://github.com/plotly/plotly.js/blob/master/dist/README.md}
}
\seealso{
\code{[plotly_bundle_info]()}
}
\author{
Carson Sievert
}
