% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{pch2symbol}
\alias{pch2symbol}
\title{Convert R pch point codes to plotly "symbol" codes.}
\format{\preformatted{ Named chr [1:20] "square" "circle" "triangle-up" "cross" "x" "diamond" ...
 - attr(*, "names")= chr [1:20] "0" "1" "2" "3" ...
}}
\usage{
pch2symbol
}
\description{
Convert R pch point codes to plotly "symbol" codes.
}
\keyword{datasets}

