% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{plot_ly}
\alias{plot_ly}
\title{Initiate a plotly visualization}
\usage{
plot_ly(data = data.frame(), ..., type = NULL, color, colors = NULL,
  symbol, symbols = NULL, size, sizes = c(10, 100), linetype,
  linetypes = NULL, width = NULL, height = NULL, source = "A")
}
\arguments{
\item{data}{A data frame (optional).}

\item{...}{These arguments are documented at \url{https://plot.ly/r/reference/}
Note that acceptable arguments depend on the value of \code{type}.}

\item{type}{A character string describing the type of trace. If \code{NULL}
(the default), the initial trace type is determined by \code{add_trace}}

\item{color}{Either a variable name or a vector to use for color mapping.}

\item{colors}{Either a colorbrewer2.org palette name (e.g. "YlOrRd" or "Blues"), 
or a vector of colors to interpolate in hexadecimal "#RRGGBB" format, 
or a color interpolation function like \code{colorRamp()}.}

\item{symbol}{Either a variable name or a (discrete) vector to use for symbol encoding.}

\item{symbols}{A character vector of symbol types. For possible values, see \link{schema}.}

\item{size}{A variable name or numeric vector to encode the size of markers.}

\item{sizes}{a numeric vector of length 2 used to scale sizes to pixels.}

\item{linetype}{Either a variable name or a (discrete) vector to use for linetype encoding.}

\item{linetypes}{A character vector of line types. For possible values, see \link{schema}.}

\item{width}{Width in pixels (optional, defaults to automatic sizing).}

\item{height}{Height in pixels (optional, defaults to automatic sizing).}

\item{source}{Only relevant for \link{event_data}.}
}
\description{
Transform data into a plotly visualization.
}
\details{
There are a number of "visual properties" that aren't included in the officical 
Reference section (see below).
}
\examples{
\dontrun{

# plot_ly() tries to create a sensible plot based on the information you 
# give it. If you don't provide a trace type, plot_ly() will infer one.
plot_ly(economics, x = ~pop)
plot_ly(economics, x = ~date, y = ~pop)
# plot_ly() doesn't require data frame(s), which allows one to take 
# advantage of trace type(s) designed specifically for numeric matrices
plot_ly(z = volcano)
plot_ly(z = volcano, type = "surface")

# plotly has a functional interface: every plotly function takes a plotly
# object as it's first input argument and returns a modified plotly object
add_points(plot_ly(economics, x = ~date, y = ~unemploy/pop))

# To make code more readable, plotly imports the pipe operator from magrittr
economics \%>\% plot_ly(x = ~date, y = ~unemploy/pop) \%>\% add_points()

# Attributes defined via plot_ly() set 'global' attributes that 
# are carried onto subsequent traces
plot_ly(economics, x = ~date, line = list(color = "black")) \%>\%
 add_trace(y = ~uempmed, mode = "markers+lines") \%>\%
 add_lines(y = ~psavert) \%>\%
 layout(title = "Setting global trace attributes")

# Attributes are documented in the figure reference -> https://plot.ly/r/reference
# You might notice plot_ly() has named arguments that aren't in the figure
# reference. These arguments make it easier to map abstract data values to
# visual attributes.
p <- plot_ly(iris, x = ~Sepal.Width, y = ~Sepal.Length) 
add_markers(p, color = ~Petal.Length)
add_markers(p, color = ~Species)
add_markers(p, color = ~Species, colors = "Set1")
add_paths(p, line = ~Species)
  
 
  
 plot_ly(economics, x = ~date) \%>\%
  add_lines(y = ~uempmed) \%>\% 
  add_lines(y = ~fitted(loess(uempmed ~ as.numeric(date)))) \%>\%
  layout(title = "Median duration of unemployment (in weeks)")

# smooth gradient between 5 colors
plot_ly(economics, x = ~date, y = ~unemploy / pop, color = ~pop, 
  colors = terrain.colors(5), mode = "markers")
  
# function to extract the decade of a given date
decade <- function(x) {
  factor(floor(as.numeric(format(x, "\%Y")) / 10) * 10)
}
plot_ly(economics, x = ~unemploy / pop, color = ~decade(date), type = "box")

plot_ly(economics, x = ~date, y = ~unemploy / pop, linetype = ~decade(date))

}

}
\author{
Carson Sievert
}
\seealso{
\code{\link{layout}()}, \code{\link{add_trace}()}, \code{\link{style}()}
}

