% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotly.R
\name{plot_ly}
\alias{plot_ly}
\title{Initiate a plotly object}
\usage{
plot_ly(data = data.frame(), type = "scatter", ..., inherit = TRUE)
}
\arguments{
\item{data}{A data frame to visualize (optional).}

\item{type}{A charater string describing the type of trace.}

\item{...}{Trace properties. See the references section below for documentation
of these properties.}

\item{inherit}{should future traces inherit properties from this initial trace?}
}
\description{
Creates a plotly object with a single trace and an empty layout. To add traces,
see \code{\link{add_trace}()} and to customize the layout see \code{\link{layout}()}.
}
\examples{
data(economics, package = "ggplot2")
# basic time-series plot
p <- plot_ly(economics, x = date, y = uempmed, type = "scatter", showlegend = F)
# add a loess smoother
p2 <- add_trace(p, y = fitted(loess(uempmed ~ as.numeric(date))))
# add a title
p3 <- layout(p2, title = "Median duration of unemployment (in weeks)")
# change the font
layout(p3, font = list(family = "Courier New, monospace"))

# sometimes, a data frame isn't fit for the use case...
# for 3D surface plots, a numeric matrix is more natural
plot_ly(z = volcano, type = "surface")
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#Trace_objects}
}

