% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotly.R
\name{plot_ly}
\alias{plot_ly}
\title{Initiate a plotly visualization}
\usage{
plot_ly(data = data.frame(), ..., type = "scatter", inherit = TRUE)
}
\arguments{
\item{data}{A data frame (optional).}

\item{...}{Visual properties.
All arguments documented in the references section below are supported.
In addition, there are special arguments which map variables to visual
aethestics in a similar style to ggplot2 (such as \code{color}).}

\item{type}{A charater string describing the type of trace.}

\item{inherit}{should future traces inherit properties from this initial trace?}
}
\description{
Transform data into a plotly visualization.
}
\examples{
data(economics, package = "ggplot2")
# basic time-series plot
p <- plot_ly(economics, x = date, y = uempmed, type = "scatter", showlegend = F)
# add a loess smoother
p2 <- add_trace(p, y = fitted(loess(uempmed ~ as.numeric(date))))
# add a title
p3 <- layout(p2, title = "Median duration of unemployment (in weeks)")
# change the font
layout(p3, font = list(family = "Courier New, monospace"))

# sometimes, a data frame isn't fit for the use case...
# for 3D surface plots, a numeric matrix is more natural
plot_ly(z = volcano, type = "surface")
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/r/reference/}
}
\seealso{
\code{\link{layout}()}, \code{\link{add_trace}()}, \code{\link{style}()}
}

