% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotly.R
\name{plotly}
\alias{plotly}
\title{Main interface to plotly}
\usage{
plotly(username = NULL, key = NULL, base_url = NULL)
}
\arguments{
\item{username}{plotly username}

\item{key}{plotly API key}

\item{base_url}{plotly server}
}
\description{
Deprecated: see \link{signup} for credentials/configuration storage details.
See \link{ggplotly} for the new ggplot2 interface.
}

