% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plotly}
\alias{plotly}
\title{Main interface to plotly}
\usage{
plotly(username, key)
}
\arguments{
\item{username}{plotly username}

\item{key}{plotly API key}
}
\description{
Deprecated: see \link{signup} for credentials/configuration storage details.
See \link{ggplotly} for the new ggplot2 interface.
}

