% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotly.R
\name{plotly}
\alias{plotly}
\title{Main interface to plotly}
\usage{
plotly(p = last_plot(), browse = interactive(), ...)
}
\arguments{
\item{p}{Either a ggplot object or a list of data/arguments to post to the
plotly api.}

\item{browse}{should the default web browser be prompted to open the Plotly result?}
}
\description{
Plotly interface object. See up-to-date documentation and examples at
https://plot.ly/API
}
\examples{
\dontrun{
# You need a plotly username and API key to communicate with
# the plotly API. These are accessed via environment variables.
# If you don't already have an API key, you can obtain one with a valid
# username and email via the signup() function.
usr <- 'anna.lyst'
Sys.setenv(`plotly-username` = usr)
resp <- signup(usr, 'anna.lyst@plot.ly')
Sys.setenv(`plotly-apikey` = resp$apikey)
# Note that you can set environment variables in your .Rprofile if you
# don't want to set them everytime you start R.

# Send data directly to Plotly's Javascript Graphing Library
# https://plot.ly/javascript-graphing-library/
p <- list(
 x = c(0, 1, 2),
 y = c(10, 11, 12)
)
resp <- plotly(p)

# plotly() also understands how to map (some) ggplot objects to Plotly graphs
ggiris <- qplot(Petal.Width, Sepal.Length, data = iris, color = Species)
plotly(ggiris)
data(canada.cities, package="maps")
viz <- ggplot(canada.cities, aes(long, lat)) +
  borders(regions="canada", name="borders") +
  coord_equal() +
  geom_point(aes(text=name, size=pop), colour="red",
               alpha=1/2, name="cities")
 plotly(viz)
}
}
\references{
https://plot.ly/API
}

