% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plotly}
\alias{plotly}
\title{Main interface to plotly}
\usage{
plotly(username, key)
}
\arguments{
\item{username}{plotly username}

\item{key}{plotly API key}
}
\description{
This function is now deprecated. It used to provide a way to store plotly
account credentials, but that can now be done with environment variables.
For more details and examples, see \url{https://plot.ly/r/getting-started/}.

If you're here looking for an intro/overview of the package, see the
\href{https://github.com/ropensci/plotly/#getting-started}{readme}
}
\seealso{
\code{\link[=ggplotly]{ggplotly()}}, \code{\link[=plot_ly]{plot_ly()}}, \code{\link[=signup]{signup()}}
}
\keyword{internal}
