% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_build.R
\name{plotly_build}
\alias{plotly_build}
\title{'Build' (i.e., evaluate) a plotly object}
\usage{
plotly_build(p)
}
\arguments{
\item{p}{a ggplot object, or a plotly object, or a list.}
}
\description{
This generic function creates the list object sent to plotly.js
for rendering. Using this function can be useful for overriding defaults
provided by \code{ggplotly}/\code{plot_ly} or for debugging rendering
errors.
}
\examples{

p <- plot_ly(economics, x = ~date, y = ~pce)
# a plotly visualization object
str(p$x$data)
# the actual list of options sent to plotly.js
str(plotly_build(p))

p <- qplot(data = mtcars, wt, mpg, geom = c("point", "smooth"))
l <- plotly_build(p)
# turn off hoverinfo for the smooth (but keep it for the points)
l$x$data[[2]]$hoverinfo <- "none"
l$x$data[[3]]$hoverinfo <- "none"
l

}

