% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotly.R
\name{plotly_build}
\alias{plotly_build}
\title{Build a plotly object before viewing it}
\usage{
plotly_build(l)
}
\arguments{
\item{l}{a ggplot object, or a plotly object, or a list.}
}
\description{
For convenience and efficiency purposes, plotly objects are subject to lazy
evaluation. That is, the actual content behind a plotly object is not
created until it is absolutely necessary. In some instances, you may want
to perform this evaluation yourself, and work directly with the resulting
list.
}

