% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_data.R
\name{plotly_data}
\alias{plotly_data}
\title{Obtain data associated with a plotly graph}
\usage{
plotly_data(p, id = p$x$cur_data)
}
\arguments{
\item{p}{a plotly visualization}

\item{id}{a character string or number referencing an "attribute layer".}
}
\value{
returns a data frame
}
\description{
\code{plotly_data()} returns data associated with 
a plotly visualization (if there are multiple data frames, by default, 
it returns the most recent one).
}

