% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_data.R
\name{plotly_data}
\alias{arrange_.plotly}
\alias{distinct_.plotly}
\alias{filter_.plotly}
\alias{group_by_.plotly}
\alias{groups.plotly}
\alias{mutate_.plotly}
\alias{plotly_data}
\alias{rename_.plotly}
\alias{select_.plotly}
\alias{slice_.plotly}
\alias{summarise_.plotly}
\alias{transmute_.plotly}
\alias{ungroup.plotly}
\title{View and manipulate data associated with a plotly graph}
\usage{
plotly_data(p, id = p$x$cur_data)

\method{groups}{plotly}(x)

\method{group_by_}{plotly}(.data, ..., .dots, add = FALSE)

\method{ungroup}{plotly}(x)

\method{summarise_}{plotly}(.data, ..., .dots)

\method{mutate_}{plotly}(.data, ..., .dots)

\method{arrange_}{plotly}(.data, ..., .dots)

\method{select_}{plotly}(.data, ..., .dots)

\method{filter_}{plotly}(.data, ..., .dots)

\method{distinct_}{plotly}(.data, ..., .dots)

\method{slice_}{plotly}(.data, ..., .dots)

\method{rename_}{plotly}(.data, ..., .dots)

\method{transmute_}{plotly}(.data, ..., .dots)
}
\description{
\code{plotly_data()} returns the data associated with 
a plotly visualization. This data can be manipulated via 
the dplyr and tidyr verbs listed below:
}
\examples{

# loading dplyr is super important!
library(dplyr)

mtcars \%>\%
  plot_ly(x = ~wt, y = ~mpg, name = "scatter trace") \%>\%
  filter(cyl == 4) \%>\%
  add_markers(x = ~wt, y = ~mpg, name = "filtered points", 
    marker = list(color = "red"))
    
library(tidyr)
economics \%>\%
  gather(variable, value, -date) \%>\%
  group_by(variable) \%>\%
  plot_ly(x = ~date, y = ~value)
  
}
\keyword{internal}

