% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotly.R
\name{plotly_POST}
\alias{plotly_POST}
\title{POST messages to plotly's REST API}
\usage{
plotly_POST(args, kwargs = list(filename = "plot from api", fileopt = "new"),
  origin = "plot", ...)
}
\arguments{
\item{args}{a list. For details see the rest API docs.}

\item{kwargs}{a list. For details see the rest API docs.}

\item{origin}{a character vector of length one. For details see the rest API docs.}

\item{...}{arguments passed along to \code{httr::POST()}}
}
\description{
POST messages to plotly's REST API
}
\examples{
args <- list(c(0, 1, 2), c(3, 4, 5), c(1, 2, 3), c(6, 6, 5))
resp <- plotly_POST(args)
}
\references{
https://plot.ly/rest/
}

