% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotly_POST.R
\name{plotly_POST}
\alias{plotly_POST}
\title{Create/Modify plotly graphs}
\usage{
plotly_POST(x)
}
\arguments{
\item{x}{either a plotly object or a list.}
}
\value{
An R object created by mapping the JSON content of the plotly API
response to its R equivalent.
}
\description{
POST messages to the clientresp resource of plotly's REST API. Unlike \link{ggplotly},
this function does not translate ggplot objects.
}
\examples{
\dontrun{
# If you want, you can still construct lists by hand...
trace1 <- list(
  x = c(1, 2, 3, 4),
  y = c(10, 15, 13, 17),
  type = "scatter"
)
trace2 <- list(
  x = c(1, 2, 3, 4),
  y = c(16, 5, 11, 9),
  type = "scatter"
)
plotly_POST(list(trace1, trace2))
}
}
\author{
Carson Sievert
}
\references{
https://plot.ly/rest/
}
\seealso{
\link{signup}
}

