% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{plotly_spec}
\alias{plotly_spec}
\title{Obtain JSON sent to plotly.js}
\usage{
plotly_spec(p = plot_ly(), ...)
}
\arguments{
\item{p}{a plotly or ggplot object.}

\item{...}{other options passed onto \code{listviewer::jsonedit}}
}
\description{
This function returns the JSON sent to plotly.js which can be useful for 
debugging.
}
\examples{
  
plotly_spec(plot_ly())
}

