% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{print.plotly}
\alias{print.plotly}
\title{Print a plotly object}
\usage{
\method{print}{plotly}(x, ...)
}
\arguments{
\item{x}{a plotly object}

\item{...}{additional arguments (currently ignored)}
}
\description{
Print a plotly object
}

