% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R, R/pipe.R
\docType{import}
\name{mutate}
\alias{mutate}
\alias{mutate_}
\alias{transmute}
\alias{transmute_}
\alias{select}
\alias{select_}
\alias{rename}
\alias{rename_}
\alias{group_by}
\alias{group_by_}
\alias{groups}
\alias{ungroup}
\alias{summarise}
\alias{summarise_}
\alias{do}
\alias{do_}
\alias{arrange}
\alias{arrange_}
\alias{distinct}
\alias{distinct_}
\alias{slice}
\alias{slice_}
\alias{filter}
\alias{filter_}
\alias{reexports}
\alias{\%>\%}
\title{Objects exported from other packages}
\usage{
mutate(.data, ...)

mutate_(.data, ..., .dots = list())

transmute(.data, ...)

transmute_(.data, ..., .dots = list())

select(.data, ...)

select_(.data, ..., .dots = list())

rename(.data, ...)

rename_(.data, ..., .dots = list())

group_by(.data, ..., .add = FALSE, .drop = group_by_drop_default(.data))

group_by_(.data, ..., .dots = list(), add = FALSE)

groups(x)

ungroup(x, ...)

summarise(.data, ..., .by = NULL, .groups = NULL)

summarise_(.data, ..., .dots = list())

do(.data, ...)

do_(.data, ..., .dots = list())

arrange(.data, ..., .by_group = FALSE)

arrange_(.data, ..., .dots = list())

distinct(.data, ..., .keep_all = FALSE)

distinct_(.data, ..., .dots, .keep_all = FALSE)

slice(.data, ..., .by = NULL, .preserve = FALSE)

slice_(.data, ..., .dots = list())

filter(.data, ..., .by = NULL, .preserve = FALSE)

filter_(.data, ..., .dots = list())
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}
}}

