% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{schema}
\alias{schema}
\title{Display plotly's plot schema}
\usage{
schema()
}
\description{
The schema contains valid attributes names, their value type, 
default values (if any), and min/max values (if applicable).
}
\examples{
schema()
}
