% Generated by roxygen2 (4.0.2): do not edit by hand
\name{show_config_file}
\alias{show_config_file}
\title{Read and print Plotly config file, wrapping get_credentials_file()}
\usage{
show_config_file(args = c())
}
\arguments{
\item{args}{Character vector of keys you are looking up}
}
\value{
List of keyword-value pairs (credentials)
}
\description{
Read and print Plotly config file, wrapping get_credentials_file()
}

