% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toRGB.R
\name{showRGB}
\alias{showRGB}
\title{View colors already formatted by toRGB()}
\usage{
showRGB(x, ...)
}
\arguments{
\item{...}{arguments passed along to \code{scales::show_col}.}

\item{A}{character string specifying color(s).}
}
\description{
Useful for viewing colors after they've been converted to plotly.js'
color format -- "rgba(255, 255, 255, 1)"
}
\examples{

showRGB(toRGB(colors()), labels = FALSE)
}
\author{
Carson Sievert
}
