% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotly.R
\name{signup}
\alias{signup}
\title{Create a new Plotly account.}
\usage{
signup(username, email)
}
\arguments{
\item{username}{Desired username}

\item{email}{Desired email}
}
\value{
\itemize{
 \item api_key key to use with the api
 \item tmp_pw temporary password to access your plotly account
}
}
\description{
A sign up interface to Plotly through the R Console.
}
\references{
https://plot.ly/rest/
}

