% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aux.R
\name{stream}
\alias{stream}
\title{Create a 'stream' auxiliary object}
\usage{
stream(token = NULL, maxpoints = NULL)
}
\arguments{
\item{token,}{maxpoints}
}
\description{
Available in \link{scatter}, \link{bar}, \link{histogram}, \link{box},
\link{heatmap}, \link{contour}, \link{histogram2d}, \link{histogram2dcontour},
\link{area}, \link{scatter3d}, \link{surface}.
}
\examples{
\dontrun{

}
}
\author{
Carson Sievert
}
\references{
\url{https://plot.ly/javascript-graphing-library/reference/#stream}
}

