% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{style}
\alias{style}
\title{Modify trace(s)}
\usage{
style(p, ..., traces = 1)
}
\arguments{
\item{p}{A plotly visualization.}

\item{...}{Visual properties.}

\item{traces}{numeric vector. Which traces should be modified?}
}
\description{
Modify trace(s) of an existing plotly visualization. Useful when used in
conjunction with \code{\link{get_figure}()}.
}
\author{
Carson Sievert
}
\seealso{
\code{\link{get_figure}()}
}

