% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotly.R
\name{style}
\alias{style}
\title{Modify trace(s)}
\usage{
style(p = get_plot(strict = FALSE), ..., traces = 1, evaluate = FALSE)
}
\arguments{
\item{p}{A plotly visualization.}

\item{...}{Visual properties.}

\item{traces}{numeric vector. Which traces should be modified?}

\item{evaluate}{logical. Evaluate arguments when this function is called?}
}
\description{
Modify trace(s) of an existing plotly visualization. Useful when used in
conjunction with \code{\link{get_figure}()}.
}
\author{
Carson Sievert
}
\seealso{
\code{\link{get_figure}()}
}

