% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/style.R
\name{style}
\alias{style}
\title{Modify trace styling}
\usage{
style(traces = 1, ...)
}
\arguments{
\item{traces}{numeric vector. Which traces should be modified?}

\item{...}{arguments coerced to a list and used to modify trace(s)}
}
\description{
Modify trace styling
}
\author{
Carson Sievert
}

