% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subplots.R
\name{subplot}
\alias{subplot}
\title{View multiple plots in a single view}
\usage{
subplot(..., nrows = 1, which_layout = 1)
}
\arguments{
\item{...}{any number of plotly objects}

\item{nrows}{number of rows for laying out plots in a grid-like structure.
Only used if no domain is specified in either \link{xaxis} or \link{yaxis}}

\item{which_layout}{adopt the layout of which plot?}
}
\value{
A plotly object
}
\description{
View multiple plots in a single view
}
\author{
Carson Sievert
}

