% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom2trace.R
\name{to_basic}
\alias{to_basic}
\title{Process layer data, mainly for 
--------------------------------------------------------------------------}
\usage{
to_basic(data, prestats_data, ...)
}
\description{
Convert geoms which are special cases of other geoms to their more "basic" 
version.
}

