% Generated by roxygen2 (4.0.2): do not edit by hand
\name{toFill}
\alias{toFill}
\title{Use default ggplot colour for fill (gray20) if not declared}
\usage{
toFill(x)
}
\arguments{
\item{x}{character for colour}
}
\value{
hexadecimal colour value
}
\description{
Use default ggplot colour for fill (gray20) if not declared
}

