% Generated by roxygen2 (4.0.2): do not edit by hand
\name{toRGB}
\alias{toRGB}
\title{Convert R colors to RGBA hexadecimal color values}
\usage{
toRGB(x, alpha = 1)
}
\arguments{
\item{x}{character for color, for example: "white"}

\item{alpha}{alpha}
}
\value{
hexadecimal color value (if is.na(x), return "none" for compatibility with JavaScript)
}
\description{
Convert R colors to RGBA hexadecimal color values
}

