% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{toWebGL}
\alias{toWebGL}
\title{Convert trace types to WebGL}
\usage{
toWebGL(p)
}
\arguments{
\item{p}{a plotly or ggplot object.}
}
\description{
Convert trace types to WebGL
}
\examples{

# currently no bargl trace type
toWebGL(ggplot() + geom_bar(aes(1:10)))
toWebGL(qplot(1:10, 1:10))

}
